Moved SimpleSplash into the widgets file. Added optional parameter "parent" in order to specify a parent widget for the splash.
