+++ .idea/workspace.xml
@@ -7,8 +7,40 @@
     <option name="autoReloadType" value="SELECTIVE" />
   </component>
   <component name="ChangeListManager">
-    <list default="true" id="952c8f52-4fbc-4a9a-a04b-64cd54717eb6" name="Changes" comment="Fixed no-chapter bible books not parsing correctly when capitalized">
+    <list default="true" id="952c8f52-4fbc-4a9a-a04b-64cd54717eb6" name="Changes" comment="Fixed no-chapter bible books being added to the OOS as 'custom_scripture'">
+      <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/docs/conf.html" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/docs/declarations.html" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/docs/edit_widget.html" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/docs/formattable_text_edit.html" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/docs/get_scripture.html" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/docs/gui.html" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/docs/help.html" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/docs/importers.html" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/docs/live_widget.html" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/docs/main.html" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/docs/media_widget.html" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/docs/oos_widget.html" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/docs/openlp_import.html" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/docs/openlyrics_export.html" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/docs/parsers.html" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/docs/preview_widget.html" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/docs/settings_widget.html" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/docs/simple_splash.html" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/docs/songselect_import.html" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/docs/toolbar.html" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/docs/web_remote.html" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/docs/widgets.html" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/edit_widget.py" beforeDir="false" afterPath="$PROJECT_DIR$/edit_widget.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/gui.py" beforeDir="false" afterPath="$PROJECT_DIR$/gui.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/importers.py" beforeDir="false" afterPath="$PROJECT_DIR$/importers.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/main.py" beforeDir="false" afterPath="$PROJECT_DIR$/main.py" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/media_widget.py" beforeDir="false" afterPath="$PROJECT_DIR$/media_widget.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/openlyrics_export.py" beforeDir="false" afterPath="$PROJECT_DIR$/openlyrics_export.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/settings_widget.py" beforeDir="false" afterPath="$PROJECT_DIR$/settings_widget.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/simple_splash.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/songselect_import.py" beforeDir="false" afterPath="$PROJECT_DIR$/songselect_import.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/widgets.py" beforeDir="false" afterPath="$PROJECT_DIR$/widgets.py" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -169,14 +201,6 @@
       <workItem from="1725650345914" duration="5458000" />
       <workItem from="1725983928582" duration="6657000" />
       <workItem from="1725995303841" duration="13965000" />
-    </task>
-    <task id="LOCAL-00089" summary="Improved the way words are drawn for songs and custom slides.&#10;&#10;Centered and/or resized pixmaps for image slides.">
-      <option name="closed" value="true" />
-      <created>1730849556485</created>
-      <option name="number" value="00089" />
-      <option name="presentableId" value="LOCAL-00089" />
-      <option name="project" value="LOCAL" />
-      <updated>1730849556486</updated>
     </task>
     <task id="LOCAL-00090" summary="Improved the way words are drawn for songs and custom slides.&#10;&#10;Centered and/or resized pixmaps for image slides.">
       <option name="closed" value="true" />
@@ -562,7 +586,15 @@
       <option name="project" value="LOCAL" />
       <updated>1757439505508</updated>
     </task>
-    <option name="localTasksCounter" value="138" />
+    <task id="LOCAL-00138" summary="Fixed no-chapter bible books being added to the OOS as 'custom_scripture'">
+      <option name="closed" value="true" />
+      <created>1757439978152</created>
+      <option name="number" value="00138" />
+      <option name="presentableId" value="LOCAL-00138" />
+      <option name="project" value="LOCAL" />
+      <updated>1757439978152</updated>
+    </task>
+    <option name="localTasksCounter" value="139" />
     <servers />
   </component>
   <component name="TypeScriptGeneratedFilesManager">
@@ -583,7 +615,6 @@
     </option>
   </component>
   <component name="VcsManagerConfiguration">
-    <MESSAGE value="Moved runnables into their own file&#10;Fixed problem with lyrics on the remote when reloading the page" />
     <MESSAGE value="Further refined the usage of blank lines when editing songs or custom slides&#10;Improved the adding and deleting of tags in the editor" />
     <MESSAGE value="Fixed editor adding a blank line to the top of text." />
     <MESSAGE value="Added ability to see release notes from About and update dialog" />
@@ -608,7 +639,8 @@
     <MESSAGE value="Added the ability to print song lyrics from the editor window&#10;    Improved printing of the Order of Service&#10;    Fixed song list updating after add, delete, or search&#10;    Ensured custom slide icons change appropriately when global backgrounds are changed&#10;    Changed font face combo boxes over to Qt's native QFontComboBox" />
     <MESSAGE value="Redesigned Font Widget using QGroupBox for each child widget." />
     <MESSAGE value="Fixed no-chapter bible books not parsing correctly when capitalized" />
-    <option name="LAST_COMMIT_MESSAGE" value="Fixed no-chapter bible books not parsing correctly when capitalized" />
+    <MESSAGE value="Fixed no-chapter bible books being added to the OOS as 'custom_scripture'" />
+    <option name="LAST_COMMIT_MESSAGE" value="Fixed no-chapter bible books being added to the OOS as 'custom_scripture'" />
   </component>
   <component name="XDebuggerManager">
     <watches-manager>
