+++ README.html
@@ -21,7 +21,7 @@ tech-savvy volunteers. ProjectOn seeks to provide an easy to use,
 intuitive, and responsive experience when building and running the
 projection needs of a church service.</p>
 
-<h1>What's New in Version 1.7.1.003?</h1>
+<h1>What's New in Version 1.7.1.004?</h1>
 <ul>
     <li>Added the ability to print song lyrics from the editor window</li>
     <li>Improved printing of the Order of Service</li>
@@ -32,7 +32,7 @@ projection needs of a church service.</p>
 
 <h1 id="installation">Installation</h1>
 <p>Currently, ProjectOn is available for the Microsoft Windows operating
-system only. Download the current ProjectOn installer (v.1.7.1.003) and run
+system only. Download the current ProjectOn installer (v.1.7.1.004) and run
 it on your computer.</p>
 
 <h1 id="using-projecton">Using ProjectOn</h1>
