+++ README.md
@@ -16,7 +16,7 @@ tech-savvy volunteers. ProjectOn seeks to provide an easy to use,
 intuitive, and responsive experience when building and running the
 projection needs of a church service.
 
-# What's New in Version 1.7.1.003?
+# What's New in Version 1.7.1.004?
 - Added the ability to print song lyrics from the editor window
 - Improved printing of the Order of Service
 - Fixed song list updating after add, delete, or search
@@ -26,7 +26,7 @@ projection needs of a church service.
 # Installation
 
 Currently, ProjectOn is available for the Microsoft Windows operating
-system only. Download the current ProjectOn installer (v.1.7.1.003) and run
+system only. Download the current ProjectOn installer (v.1.7.1.004) and run
 it on your computer.
 
 # Using ProjectOn
