--- docs/declarations.html
@@ -1,24 +1,0 @@
-<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.0 Transitional//EN">
-<html><head><title>Python: module declarations</title>
-<meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-</head><body bgcolor="#f0f0f8">
-
-<table width="100%" cellspacing=0 cellpadding=2 border=0 summary="heading">
-<tr bgcolor="#7799ee">
-<td valign=bottom>&nbsp;<br>
-<font color="#ffffff" face="helvetica, arial">&nbsp;<br><big><big><strong>declarations</strong></big></big></font></td
-><td align=right valign=bottom
-><font color="#ffffff" face="helvetica, arial"><a href=".">index</a><br><a href="file:/home/jeremy/Documents/Python%20Workspace/project_on/declarations.py">/home/jeremy/Documents/Python Workspace/project_on/declarations.py</a></font></td></tr></table>
-    <p></p>
-<p>
-<table width="100%" cellspacing=0 cellpadding=2 border=0 summary="section">
-<tr bgcolor="#55aa55">
-<td colspan=3 valign=bottom>&nbsp;<br>
-<font color="#ffffff" face="helvetica, arial"><big><strong>Data</strong></big></font></td></tr>
-    
-<tr><td bgcolor="#55aa55"><tt>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;</tt></td><td>&nbsp;</td>
-<td width="100%"><strong>DB_STRUCTURE</strong> = {'backgroundThumbnails': {'fileName': 'TEXT', 'image': 'BLOB'}, 'customSlides': {'audio_file': 'TEXT', 'auto_play': 'TEXT', 'background': 'TEXT', 'font': 'TEXT', 'fontColor': 'TEXT', 'font_size': 'TEXT', 'loop_audio': 'TEXT', 'outline_color': 'TEXT', 'outline_width': 'TEXT', 'override_global': 'TEXT', ...}, 'imageThumbnails': {'fileName': 'TEXT', 'image': 'BLOB'}, 'songs': {'author': 'TEXT', 'background': 'TEXT', 'ccliNum': 'TEXT', 'copyright': 'TEXT', 'font': 'TEXT', 'fontColor': 'TEXT', 'font_size': 'TEXT', 'footer': 'TEXT', 'lyrics': 'TEXT', 'outline_color': 'TEXT', ...}, 'web': {'title': 'TEXT', 'url': 'TEXT'}}<br>
-<strong>SLIDE_DATA_DEFAULTS</strong> = {'audio_file': '', 'author': '', 'auto_play': False, 'background': '', 'ccli_song_number': '', 'copyright': '', 'file_name': '', 'font_color': 'black', 'font_family': 'Arial', 'font_size': 72, ...}<br>
-<strong>SQL_COLUMN_TO_DICTIONARY_CUSTOM</strong> = {0: 'title', 1: 'text', 2: 'font', 3: 'font_color', 4: 'background', 5: 'font_size', 6: 'use_shadow', 7: 'shadow_color', 8: 'shadow_offset', 9: 'use_outline', ...}<br>
-<strong>SQL_COLUMN_TO_DICTIONARY_SONG</strong> = {0: 'title', 1: 'author', 2: 'copyright', 3: 'ccli_song_number', 4: 'text', 5: 'verse_order', 6: 'footer', 7: 'font_family', 8: 'font_color', 9: 'background', ...}</td></tr></table>
-</body></html>
