--- docs/get_scripture.html
@@ -1,70 +1,0 @@
-<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.0 Transitional//EN">
-<html><head><title>Python: module get_scripture</title>
-<meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-</head><body bgcolor="#f0f0f8">
-
-<table width="100%" cellspacing=0 cellpadding=2 border=0 summary="heading">
-<tr bgcolor="#7799ee">
-<td valign=bottom>&nbsp;<br>
-<font color="#ffffff" face="helvetica, arial">&nbsp;<br><big><big><strong>get_scripture</strong></big></big></font></td
-><td align=right valign=bottom
-><font color="#ffffff" face="helvetica, arial"><a href=".">index</a><br><a href="file:/home/jeremy/Documents/Python%20Workspace/project_on/get_scripture.py">/home/jeremy/Documents/Python Workspace/project_on/get_scripture.py</a></font></td></tr></table>
-    <p></p>
-<p>
-<table width="100%" cellspacing=0 cellpadding=2 border=0 summary="section">
-<tr bgcolor="#aa55cc">
-<td colspan=3 valign=bottom>&nbsp;<br>
-<font color="#ffffff" face="helvetica, arial"><big><strong>Modules</strong></big></font></td></tr>
-    
-<tr><td bgcolor="#aa55cc"><tt>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;</tt></td><td>&nbsp;</td>
-<td width="100%"><table width="100%" summary="list"><tr><td width="25%" valign=top><a href="xml.etree.ElementTree.html">xml.etree.ElementTree</a><br>
-</td><td width="25%" valign=top><a href="re.html">re</a><br>
-</td><td width="25%" valign=top></td><td width="25%" valign=top></td></tr></table></td></tr></table><p>
-<table width="100%" cellspacing=0 cellpadding=2 border=0 summary="section">
-<tr bgcolor="#ee77aa">
-<td colspan=3 valign=bottom>&nbsp;<br>
-<font color="#ffffff" face="helvetica, arial"><big><strong>Classes</strong></big></font></td></tr>
-    
-<tr><td bgcolor="#ee77aa"><tt>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;</tt></td><td>&nbsp;</td>
-<td width="100%"><dl>
-<dt><font face="helvetica, arial"><a href="builtins.html#object">builtins.object</a>
-</font></dt><dd>
-<dl>
-<dt><font face="helvetica, arial"><a href="get_scripture.html#GetScripture">GetScripture</a>
-</font></dt></dl>
-</dd>
-</dl>
- <p>
-<table width="100%" cellspacing=0 cellpadding=2 border=0 summary="section">
-<tr bgcolor="#ffc8d8">
-<td colspan=3 valign=bottom>&nbsp;<br>
-<font color="#000000" face="helvetica, arial"><a name="GetScripture">class <strong>GetScripture</strong></a>(<a href="builtins.html#object">builtins.object</a>)</font></td></tr>
-    
-<tr bgcolor="#ffc8d8"><td rowspan=2><tt>&nbsp;&nbsp;&nbsp;</tt></td>
-<td colspan=2><tt><a href="#GetScripture">GetScripture</a>(main)<br>
-&nbsp;<br>
-<a href="#GetScripture">GetScripture</a>&nbsp;is&nbsp;a&nbsp;class&nbsp;that&nbsp;will&nbsp;retrieve&nbsp;a&nbsp;specific&nbsp;scripture&nbsp;passage&nbsp;from&nbsp;the&nbsp;user's&nbsp;xml&nbsp;bible&nbsp;based&nbsp;on<br>
-what&nbsp;is&nbsp;typed&nbsp;in&nbsp;the&nbsp;Scripture&nbsp;Reference&nbsp;LineEdit.&nbsp;The&nbsp;xml&nbsp;bible&nbsp;is&nbsp;expected&nbsp;to&nbsp;be&nbsp;in&nbsp;the&nbsp;Zefania&nbsp;XML&nbsp;Bible&nbsp;format.<br>&nbsp;</tt></td></tr>
-<tr><td>&nbsp;</td>
-<td width="100%">Methods defined here:<br>
-<dl><dt><a name="GetScripture-__init__"><strong>__init__</strong></a>(self, main)</dt><dd><tt><a href="#GetScripture">GetScripture</a>&nbsp;is&nbsp;a&nbsp;class&nbsp;that&nbsp;will&nbsp;retrieve&nbsp;a&nbsp;specific&nbsp;scripture&nbsp;passage&nbsp;from&nbsp;the&nbsp;user's&nbsp;xml&nbsp;bible&nbsp;based&nbsp;on<br>
-what&nbsp;is&nbsp;typed&nbsp;in&nbsp;the&nbsp;Scripture&nbsp;Reference&nbsp;LineEdit.&nbsp;The&nbsp;xml&nbsp;bible&nbsp;is&nbsp;expected&nbsp;to&nbsp;be&nbsp;in&nbsp;the&nbsp;Zefania&nbsp;XML&nbsp;Bible&nbsp;format.<br>
-:param&nbsp;ProjectOn&nbsp;main:&nbsp;The&nbsp;current&nbsp;instance&nbsp;of&nbsp;ProjectOn</tt></dd></dl>
-
-<dl><dt><a name="GetScripture-get_passage"><strong>get_passage</strong></a>(self, reference)</dt><dd><tt>Method&nbsp;to&nbsp;parse&nbsp;the&nbsp;user's&nbsp;inputted&nbsp;reference&nbsp;and&nbsp;retrieve&nbsp;the&nbsp;passage&nbsp;from&nbsp;the&nbsp;user's&nbsp;xml&nbsp;bible.<br>
-:param&nbsp;str&nbsp;reference:&nbsp;The&nbsp;user-provided&nbsp;scripture&nbsp;reference</tt></dd></dl>
-
-<hr>
-Data descriptors defined here:<br>
-<dl><dt><strong>__dict__</strong></dt>
-<dd><tt>dictionary&nbsp;for&nbsp;instance&nbsp;variables&nbsp;(if&nbsp;defined)</tt></dd>
-</dl>
-<dl><dt><strong>__weakref__</strong></dt>
-<dd><tt>list&nbsp;of&nbsp;weak&nbsp;references&nbsp;to&nbsp;the&nbsp;object&nbsp;(if&nbsp;defined)</tt></dd>
-</dl>
-<hr>
-Data and other attributes defined here:<br>
-<dl><dt><strong>root</strong> = None</dl>
-
-</td></tr></table></td></tr></table>
-</body></html>
