--- docs/openlp_import.html
@@ -1,74 +1,0 @@
-<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.0 Transitional//EN">
-<html><head><title>Python: module openlp_import</title>
-<meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-</head><body bgcolor="#f0f0f8">
-
-<table width="100%" cellspacing=0 cellpadding=2 border=0 summary="heading">
-<tr bgcolor="#7799ee">
-<td valign=bottom>&nbsp;<br>
-<font color="#ffffff" face="helvetica, arial">&nbsp;<br><big><big><strong>openlp_import</strong></big></big></font></td
-><td align=right valign=bottom
-><font color="#ffffff" face="helvetica, arial"><a href=".">index</a><br><a href="file:/home/jeremy/Documents/Python%20Workspace/project_on/openlp_import.py">/home/jeremy/Documents/Python Workspace/project_on/openlp_import.py</a></font></td></tr></table>
-    <p></p>
-<p>
-<table width="100%" cellspacing=0 cellpadding=2 border=0 summary="section">
-<tr bgcolor="#aa55cc">
-<td colspan=3 valign=bottom>&nbsp;<br>
-<font color="#ffffff" face="helvetica, arial"><big><strong>Modules</strong></big></font></td></tr>
-    
-<tr><td bgcolor="#aa55cc"><tt>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;</tt></td><td>&nbsp;</td>
-<td width="100%"><table width="100%" summary="list"><tr><td width="25%" valign=top><a href="xml.etree.ElementTree.html">xml.etree.ElementTree</a><br>
-</td><td width="25%" valign=top><a href="os.html">os</a><br>
-</td><td width="25%" valign=top><a href="re.html">re</a><br>
-</td><td width="25%" valign=top><a href="sqlite3.html">sqlite3</a><br>
-</td></tr></table></td></tr></table><p>
-<table width="100%" cellspacing=0 cellpadding=2 border=0 summary="section">
-<tr bgcolor="#ee77aa">
-<td colspan=3 valign=bottom>&nbsp;<br>
-<font color="#ffffff" face="helvetica, arial"><big><strong>Classes</strong></big></font></td></tr>
-    
-<tr><td bgcolor="#ee77aa"><tt>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;</tt></td><td>&nbsp;</td>
-<td width="100%"><dl>
-<dt><font face="helvetica, arial"><a href="builtins.html#object">builtins.object</a>
-</font></dt><dd>
-<dl>
-<dt><font face="helvetica, arial"><a href="openlp_import.html#OpenLPImport">OpenLPImport</a>
-</font></dt></dl>
-</dd>
-</dl>
- <p>
-<table width="100%" cellspacing=0 cellpadding=2 border=0 summary="section">
-<tr bgcolor="#ffc8d8">
-<td colspan=3 valign=bottom>&nbsp;<br>
-<font color="#000000" face="helvetica, arial"><a name="OpenLPImport">class <strong>OpenLPImport</strong></a>(<a href="builtins.html#object">builtins.object</a>)</font></td></tr>
-    
-<tr bgcolor="#ffc8d8"><td rowspan=2><tt>&nbsp;&nbsp;&nbsp;</tt></td>
-<td colspan=2><tt><a href="#OpenLPImport">OpenLPImport</a>(gui)<br>
-&nbsp;<br>
-Provides&nbsp;the&nbsp;ability&nbsp;to&nbsp;import&nbsp;and&nbsp;parse&nbsp;songs&nbsp;from&nbsp;an&nbsp;OpenLP&nbsp;sqlite&nbsp;database.<br>&nbsp;</tt></td></tr>
-<tr><td>&nbsp;</td>
-<td width="100%">Methods defined here:<br>
-<dl><dt><a name="OpenLPImport-__init__"><strong>__init__</strong></a>(self, gui)</dt><dd><tt>Provides&nbsp;the&nbsp;ability&nbsp;to&nbsp;import&nbsp;and&nbsp;parse&nbsp;songs&nbsp;from&nbsp;an&nbsp;OpenLP&nbsp;sqlite&nbsp;database.<br>
-:param&nbsp;gui.GUI&nbsp;gui:&nbsp;the&nbsp;current&nbsp;instance&nbsp;of&nbsp;GUI</tt></dd></dl>
-
-<dl><dt><a name="OpenLPImport-convert_lyrics"><strong>convert_lyrics</strong></a>(self, lyrics)</dt><dd><tt>Method&nbsp;to&nbsp;change&nbsp;OpenLp's&nbsp;segment&nbsp;tags&nbsp;to&nbsp;this&nbsp;program's&nbsp;segment&nbsp;tags.<br>
-:param&nbsp;str&nbsp;lyrics:&nbsp;The&nbsp;song's&nbsp;lyrics<br>
-:return&nbsp;str&nbsp;converted_lyrics:&nbsp;The&nbsp;reformatted&nbsp;lyrics</tt></dd></dl>
-
-<dl><dt><a name="OpenLPImport-init_components"><strong>init_components</strong></a>(self)</dt><dd><tt>Creates&nbsp;and&nbsp;lays&nbsp;out&nbsp;the&nbsp;widget&nbsp;prompting&nbsp;the&nbsp;user&nbsp;to&nbsp;import&nbsp;their&nbsp;database&nbsp;and&nbsp;shows&nbsp;a&nbsp;progress&nbsp;bar&nbsp;when<br>
-importing.</tt></dd></dl>
-
-<dl><dt><a name="OpenLPImport-load_file"><strong>load_file</strong></a>(self)</dt><dd><tt>Creates&nbsp;a&nbsp;QFileDialog&nbsp;for&nbsp;the&nbsp;user&nbsp;to&nbsp;locate&nbsp;their&nbsp;OpenLP&nbsp;Database</tt></dd></dl>
-
-<dl><dt><a name="OpenLPImport-start_import"><strong>start_import</strong></a>(self)</dt><dd><tt>Method&nbsp;to&nbsp;perform&nbsp;the&nbsp;task&nbsp;of&nbsp;retrieving&nbsp;the&nbsp;song&nbsp;data&nbsp;from&nbsp;the&nbsp;OpenLP&nbsp;database.</tt></dd></dl>
-
-<hr>
-Data descriptors defined here:<br>
-<dl><dt><strong>__dict__</strong></dt>
-<dd><tt>dictionary&nbsp;for&nbsp;instance&nbsp;variables&nbsp;(if&nbsp;defined)</tt></dd>
-</dl>
-<dl><dt><strong>__weakref__</strong></dt>
-<dd><tt>list&nbsp;of&nbsp;weak&nbsp;references&nbsp;to&nbsp;the&nbsp;object&nbsp;(if&nbsp;defined)</tt></dd>
-</dl>
-</td></tr></table></td></tr></table>
-</body></html>
