--- docs/parsers.html
@@ -1,82 +1,0 @@
-<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.0 Transitional//EN">
-<html><head><title>Python: module parsers</title>
-<meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-</head><body bgcolor="#f0f0f8">
-
-<table width="100%" cellspacing=0 cellpadding=2 border=0 summary="heading">
-<tr bgcolor="#7799ee">
-<td valign=bottom>&nbsp;<br>
-<font color="#ffffff" face="helvetica, arial">&nbsp;<br><big><big><strong>parsers</strong></big></big></font></td
-><td align=right valign=bottom
-><font color="#ffffff" face="helvetica, arial"><a href=".">index</a><br><a href="file:/home/jeremy/Documents/Python%20Workspace/project_on/parsers.py">/home/jeremy/Documents/Python Workspace/project_on/parsers.py</a></font></td></tr></table>
-    <p></p>
-<p>
-<table width="100%" cellspacing=0 cellpadding=2 border=0 summary="section">
-<tr bgcolor="#aa55cc">
-<td colspan=3 valign=bottom>&nbsp;<br>
-<font color="#ffffff" face="helvetica, arial"><big><strong>Modules</strong></big></font></td></tr>
-    
-<tr><td bgcolor="#aa55cc"><tt>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;</tt></td><td>&nbsp;</td>
-<td width="100%"><table width="100%" summary="list"><tr><td width="25%" valign=top><a href="re.html">re</a><br>
-</td><td width="25%" valign=top></td><td width="25%" valign=top></td><td width="25%" valign=top></td></tr></table></td></tr></table><p>
-<table width="100%" cellspacing=0 cellpadding=2 border=0 summary="section">
-<tr bgcolor="#ee77aa">
-<td colspan=3 valign=bottom>&nbsp;<br>
-<font color="#ffffff" face="helvetica, arial"><big><strong>Classes</strong></big></font></td></tr>
-    
-<tr><td bgcolor="#ee77aa"><tt>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;</tt></td><td>&nbsp;</td>
-<td width="100%"><dl>
-<dt><font face="helvetica, arial"><a href="builtins.html#object">builtins.object</a>
-</font></dt><dd>
-<dl>
-<dt><font face="helvetica, arial"><a href="parsers.html#ParseScriptureReference">ParseScriptureReference</a>
-</font></dt></dl>
-</dd>
-</dl>
- <p>
-<table width="100%" cellspacing=0 cellpadding=2 border=0 summary="section">
-<tr bgcolor="#ffc8d8">
-<td colspan=3 valign=bottom>&nbsp;<br>
-<font color="#000000" face="helvetica, arial"><a name="ParseScriptureReference">class <strong>ParseScriptureReference</strong></a>(<a href="builtins.html#object">builtins.object</a>)</font></td></tr>
-    
-<tr bgcolor="#ffc8d8"><td rowspan=2><tt>&nbsp;&nbsp;&nbsp;</tt></td>
-<td colspan=2><tt>Class&nbsp;to&nbsp;take&nbsp;a&nbsp;human-readable&nbsp;scripture&nbsp;reference&nbsp;and&nbsp;split/standardize&nbsp;it&nbsp;according&nbsp;to&nbsp;book,&nbsp;chapter(s)&nbsp;and&nbsp;verses<br>&nbsp;</tt></td></tr>
-<tr><td>&nbsp;</td>
-<td width="100%">Methods defined here:<br>
-<dl><dt><a name="ParseScriptureReference-__init__"><strong>__init__</strong></a>(self)</dt><dd><tt>Class&nbsp;to&nbsp;take&nbsp;a&nbsp;human-readable&nbsp;scripture&nbsp;reference&nbsp;and&nbsp;split/standardize&nbsp;it&nbsp;according&nbsp;to&nbsp;book,&nbsp;chapter(s)&nbsp;and&nbsp;verses</tt></dd></dl>
-
-<dl><dt><a name="ParseScriptureReference-parse_reference"><strong>parse_reference</strong></a>(self, reference)</dt><dd><tt>Provides&nbsp;a&nbsp;method&nbsp;to&nbsp;parse&nbsp;a&nbsp;human-readable&nbsp;scripture&nbsp;reference&nbsp;into&nbsp;its&nbsp;book,&nbsp;chapter(s)&nbsp;and&nbsp;verse(s),<br>
-comparing&nbsp;the&nbsp;book&nbsp;to&nbsp;the&nbsp;commonly-used&nbsp;abbreviations&nbsp;defined&nbsp;in&nbsp;__init__&nbsp;and&nbsp;determining&nbsp;if&nbsp;the&nbsp;book<br>
-has&nbsp;no&nbsp;chapters.&nbsp;Returns&nbsp;a&nbsp;dictionary&nbsp;containing&nbsp;book,&nbsp;chapter_start,&nbsp;chapter_end,&nbsp;verse_start,&nbsp;verse_end,<br>
-and&nbsp;a&nbsp;boolean&nbsp;stating&nbsp;whether&nbsp;the&nbsp;reference&nbsp;contains&nbsp;a&nbsp;book&nbsp;name&nbsp;that&nbsp;can&nbsp;be&nbsp;standardized.<br>
-:param&nbsp;str&nbsp;reference:&nbsp;the&nbsp;scripture&nbsp;reference<br>
-:return:&nbsp;dict</tt></dd></dl>
-
-<hr>
-Data descriptors defined here:<br>
-<dl><dt><strong>__dict__</strong></dt>
-<dd><tt>dictionary&nbsp;for&nbsp;instance&nbsp;variables&nbsp;(if&nbsp;defined)</tt></dd>
-</dl>
-<dl><dt><strong>__weakref__</strong></dt>
-<dd><tt>list&nbsp;of&nbsp;weak&nbsp;references&nbsp;to&nbsp;the&nbsp;object&nbsp;(if&nbsp;defined)</tt></dd>
-</dl>
-</td></tr></table></td></tr></table><p>
-<table width="100%" cellspacing=0 cellpadding=2 border=0 summary="section">
-<tr bgcolor="#eeaa77">
-<td colspan=3 valign=bottom>&nbsp;<br>
-<font color="#ffffff" face="helvetica, arial"><big><strong>Functions</strong></big></font></td></tr>
-    
-<tr><td bgcolor="#eeaa77"><tt>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;</tt></td><td>&nbsp;</td>
-<td width="100%"><dl><dt><a name="-parse_scripture_by_verse"><strong>parse_scripture_by_verse</strong></a>(gui, text)</dt><dd><tt>Take&nbsp;a&nbsp;passage&nbsp;of&nbsp;scripture&nbsp;and&nbsp;split&nbsp;it&nbsp;according&nbsp;to&nbsp;how&nbsp;many&nbsp;verses&nbsp;will&nbsp;fit&nbsp;on&nbsp;the&nbsp;display&nbsp;screen,&nbsp;given<br>
-the&nbsp;current&nbsp;font&nbsp;and&nbsp;size.<br>
-:param&nbsp;GUI&nbsp;gui:&nbsp;The&nbsp;current&nbsp;instance&nbsp;of&nbsp;GUI<br>
-:param&nbsp;list&nbsp;of&nbsp;str&nbsp;text:&nbsp;The&nbsp;bible&nbsp;passage&nbsp;to&nbsp;be&nbsp;split</tt></dd></dl>
- <dl><dt><a name="-parse_scripture_item"><strong>parse_scripture_item</strong></a>(gui, text)</dt><dd><tt>Method&nbsp;to&nbsp;take&nbsp;a&nbsp;scripture&nbsp;passage&nbsp;and&nbsp;divide&nbsp;it&nbsp;up&nbsp;according&nbsp;to&nbsp;what&nbsp;will&nbsp;fit&nbsp;on&nbsp;the&nbsp;screen&nbsp;given&nbsp;the&nbsp;current<br>
-font&nbsp;and&nbsp;size.<br>
-:param&nbsp;GUI&nbsp;gui:&nbsp;The&nbsp;current&nbsp;instance&nbsp;of&nbsp;GUI<br>
-:param&nbsp;str&nbsp;text:&nbsp;The&nbsp;scripture&nbsp;text&nbsp;to&nbsp;be&nbsp;parsed</tt></dd></dl>
- <dl><dt><a name="-parse_song_data"><strong>parse_song_data</strong></a>(gui, song_data)</dt><dd><tt>Method&nbsp;to&nbsp;take&nbsp;the&nbsp;stored&nbsp;lyrics&nbsp;of&nbsp;a&nbsp;song&nbsp;and&nbsp;parse&nbsp;them&nbsp;out&nbsp;according&nbsp;to&nbsp;their&nbsp;segment&nbsp;markers&nbsp;(i.e.&nbsp;[V1])<br>
-:param&nbsp;GUI&nbsp;gui:&nbsp;The&nbsp;current&nbsp;instance&nbsp;of&nbsp;GUI<br>
-:param&nbsp;str&nbsp;lyrics:&nbsp;The&nbsp;raw&nbsp;lyrics&nbsp;data</tt></dd></dl>
-</td></tr></table>
-</body></html>
