+++ gui.py
@@ -31,7 +31,7 @@ from openlyrics_export import OpenlyricsExport
 from preview_widget import PreviewWidget
 from runnables import TimedPreviewUpdate, SlideAutoPlay, CountdownTimer
 from settings_widget import SettingsWidget
-from simple_splash import SimpleSplash
+from widgets import SimpleSplash
 from songselect_import import SongselectImport
 from toolbar import Toolbar
 from widgets import CustomMainWindow, DisplayWidget, LyricDisplayWidget, StandardItemWidget, CountdownWidget
@@ -648,7 +648,7 @@ class GUI(QObject):
             wait_widget.widget.deleteLater()
 
     def check_update(self):
-        current_version = 'v.1.7.1.003'
+        current_version = 'v.1.7.1.004'
         current_version = current_version.replace('v.', '')
         current_version = current_version.replace('rc', '')
         current_version_split = current_version.split('.')
@@ -968,7 +968,7 @@ class GUI(QObject):
         title_pixmap_label.setPixmap(title_pixmap)
         title_widget.layout().addWidget(title_pixmap_label)
 
-        title_label = QLabel('ProjectOn v.1.7.1.003')
+        title_label = QLabel('ProjectOn v.1.7.1.004')
         title_label.setFont(QFont('Helvetica', 24, QFont.Weight.Bold))
         title_widget.layout().addWidget(title_label)
         title_widget.layout().addStretch()
