+++ widgets.py
@@ -13,340 +13,117 @@ from PyQt5.QtWidgets import QListWidget, QLabel, QListWidgetItem, QComboBox, QLi
     QApplication, QFontComboBox, QGroupBox
 
 
-class FontFaceListWidget(QListWidget):
+class AutoSelectLineEdit(QLineEdit):
     """
-    Creates a custom QListWidget that displays all fonts on the system in their own style.
-    :param gui.GUI gui: The current instance of GUI
+    Implements QLineEdit to add the ability to select all text when this line edit receives focus.
     """
-    def __init__(self, gui):
-        """
-        :param gui.GUI gui: The current instance of GUI
-        """
+    def __init__(self):
         super().__init__()
-        self.gui = gui
-        self.setObjectName('FontFaceListWidget')
-        self.setMinimumHeight(60)
-        self.blockSignals(True)
-        self.populate_widget()
-        self.blockSignals(False)
 
-    def populate_widget(self):
-        try:
-            families = QFontDatabase().families()
-            for font in families:
-                if self.gui.main.initial_startup:
-                    self.gui.main.update_status_signal.emit('Processing Fonts', 'status')
-                    self.gui.main.update_status_signal.emit(font, 'info')
-                list_label = QLabel(font)
-                list_label.setFont(QFont(font, 12))
-                item = QListWidgetItem()
-                item.setData(20, font)
-                self.addItem(item)
-                self.setItemWidget(item, list_label)
-
-            if self.gui.main.initial_startup:
-                self.gui.main.update_status_signal.emit('', 'info')
-        except Exception:
-            self.gui.main.error_log()
+    def focusInEvent(self, evt):
+        super().focusInEvent(evt)
+        QTimer.singleShot(0, self.selectAll)
 
 
-class FontFaceComboBox(QComboBox):
-    """
-    Creates a custom QComboBox that displays all fonts on the system in their own style.
-    :param gui.GUI gui: The current instance of GUI
-    """
-    current_font = None
-
-    def __init__(self, gui):
-        """
-        :param gui.GUI gui: The current instance of GUI
-        """
-        super().__init__()
-        self.setObjectName('FontFaceComboBox')
-        self.gui = gui
-        self.setEditable(True)
-        self.setSizeAdjustPolicy(QComboBox.AdjustToMinimumContentsLengthWithIcon)
-        self.populate_widget()
-
-    def populate_widget(self):
-        try:
-            for i in range(len(self.gui.font_pixmaps)):
-                if i == len(self.gui.font_pixmaps) - 1:
-                    self.setIconSize(QSize(self.gui.font_pixmaps[i][0], self.gui.font_pixmaps[i][1]))
-                    self.setMinimumWidth(self.gui.font_pixmaps[i][0])
-                else:
-                    self.addItem(QIcon(self.gui.font_pixmaps[i][1]), self.gui.font_pixmaps[i][0])
-        except Exception:
-            self.gui.main.error_log()
-
-        for i in range(self.count()):
-            if self.itemText(i) == self.current_font:
-                self.setCurrentIndex(i)
-                break
-
-    def wheelEvent(self, evt):
-        evt.ignore()
-
-
-class ImageCombobox(QComboBox):
-    """
-    Creates a custom QComboBox that displays a thumbnail of an image to be used.
-    """
-    def __init__(self, gui, type, suppress_autosave=False):
-        """
-        :param gui.GUI gui: The current instance of GUI
-        :param str type: Whether this is creating a combobox of 'logo', 'song', or 'bible' images
-        """
-        super().__init__()
-        self.gui = gui
-        self.type = type
-        self.table = None
-        self.suppress_autosave = suppress_autosave
-        self.setView(QListView())
-        self.setObjectName(type)
-
-        self.setIconSize(QSize(96, 54))
-        self.setMaximumWidth(240)
-        self.setFont(self.gui.standard_font)
-
-        self.currentIndexChanged.connect(self.index_changed)
-
-        if type == 'edit_background':
-            self.removeItem(0)
-        else:
-            self.currentIndexChanged.connect(self.gui.tool_bar.change_background)
-        self.refresh()
-
-    def index_changed(self):
-        file_name = self.itemData(self.currentIndex(), Qt.ItemDataRole.UserRole)
-        if not file_name:
-            return
-        if self.type == 'logo':
-            self.gui.main.settings['logo_image'] = file_name
-        elif self.type == 'song':
-            self.gui.main.settings['global_song_background'] = file_name
-            self.gui.global_song_background_pixmap = QPixmap(self.gui.main.image_dir + '/' + file_name)
-        elif self.type == 'bible':
-            self.gui.main.settings['global_bible_background'] = file_name
-            self.gui.global_bible_background_pixmap = QPixmap(self.gui.main.image_dir + '/' + file_name)
-
-        for i in range(self.gui.oos_widget.oos_list_widget.count()):
-            item = self.gui.oos_widget.oos_list_widget.item(i)
-            item_data = item.data(Qt.ItemDataRole.UserRole)
-            pixmap = None
-
-            if item_data['type'] == 'song' or item_data['type'] == 'bible' or item_data['type'] == 'custom':
-                if not item_data['override_global'] or item_data['override_global'] == 'False':
-                    if item_data['type'] == self.type:
-                        pixmap = self.itemIcon(self.currentIndex()).pixmap(QSize(50, 27))
-                    elif item_data['type'] == 'custom' and self.type == 'bible':
-                        pixmap = self.itemIcon(self.currentIndex()).pixmap(QSize(50, 27))
-                else:
-                    if item_data['background'] == 'global_song' and self.type == 'song':
-                        pixmap = self.itemIcon(self.currentIndex()).pixmap(QSize(50, 27))
-                    elif item_data['background'] == 'global_bible' and self.type == 'bible':
-                        pixmap = self.itemIcon(self.currentIndex()).pixmap(QSize(50, 27))
-
-            if pixmap:
-                widget = self.gui.oos_widget.oos_list_widget.itemWidget(item)
-                widget.icon.setPixmap(pixmap)
-                widget.adjustSize()
-                item.setSizeHint(widget.sizeHint())
-
-        if not self.suppress_autosave:
-            self.gui.main.save_settings()
-
-    def refresh(self):
-        """
-        Method to refresh the combo box after changes to the image indices
-        """
-        self.blockSignals(True)
-        self.clear()
-
-        if self.type == 'logo':
-            self.addItem('Choose Logo Image', userData='choose_logo')
-            self.addItem('Import a Logo Image', userData='import_logo')
-            self.table = 'imageThumbnails'
-        elif self.type == 'edit':
-            self.addItem('Choose Custom Background', userData='choose_global')
-            self.table = 'backgroundThumbnails'
-        elif self.type == 'delete_background':
-            self.addItem('Choose Background to Remove')
-            self.table = 'backgroundThumbnails'
-        elif self.type == 'delete_image':
-            self.addItem('Choose Image Item to Remove')
-            self.table = 'imageThumbnails'
-        else:
-            self.addItem('Choose Global ' + self.type + ' Background', userData='choose_global')
-            self.addItem('Import a Background Image', userData='import_global')
-            self.table = 'backgroundThumbnails'
-        connection = None
-
-        try:
-            image_list = []
-            connection = sqlite3.connect(self.gui.main.database)
-            cursor = connection.cursor()
-            thumbnails = cursor.execute(
-                'SELECT * FROM ' + self.table + ' ORDER BY fileName COLLATE NOCASE ASC;').fetchall()
-            for record in thumbnails:
-                if self.gui.main.initial_startup:
-                    self.gui.main.update_status_signal.emit('Loading Thumbnails', 'status')
-                    self.gui.main.update_status_signal.emit(record[0], 'info')
-                pixmap = QPixmap()
-                pixmap.loadFromData(record[1], 'JPG')
-                icon = QIcon(pixmap)
-                self.addItem(icon, record[0].split('.')[0], userData=record[0])
-                image_list.append([icon, record[0].split('.')[0], record[0]])
-            connection.close()
-
-            if self.gui.main.initial_startup:
-                self.gui.main.update_status_signal.emit('', 'info')
-
-            self.blockSignals(False)
-        except Exception:
-            self.gui.main.error_log()
-            if connection:
-                connection.close()
-            self.blockSignals(False)
-
-    def wheelEvent(self, evt):
-        # prevent wheel scrolling, which is undesirable in the settings layout
-        evt.ignore()
-
-
-class ShadowSlider(QWidget):
-    """
-    Creates a widget containing a QSlider and Label which lets the user set the greyness of the display's shadow
-    :param gui.GUI gui: The current instance of GUI
-    """
-    def __init__(self, gui):
-        """
-        Creates a widget containing a QSlider and Label which lets the user set the greyness of the display's shadow
-        :param gui.GUI gui: The current instance of GUI
-        """
-        super().__init__()
-        self.gui = gui
-        layout = QVBoxLayout()
-        layout.setContentsMargins(0, 0, 0, 0)
-        layout.setSpacing(0)
-        self.setLayout(layout)
-
-        self.color_title = QLabel('Shadow Shade:')
-        self.color_title.setFont(self.gui.list_font)
-        layout.addWidget(self.color_title)
-
-        slider_widget = QWidget()
-        slider_widget.setFixedWidth(300)
-        slider_layout = QGridLayout()
-        slider_layout.setContentsMargins(0, 0, 0, 0)
-        slider_layout.setVerticalSpacing(0)
-        slider_widget.setLayout(slider_layout)
-        layout.addWidget(slider_widget)
-
-        self.color_slider = CustomSlider()
-        self.color_slider.setObjectName('color_slider')
-        self.color_slider.setOrientation(Qt.Orientation.Horizontal)
-        self.color_slider.setFont(self.gui.list_font)
-        self.color_slider.setRange(0, 255)
-        self.color_slider.installEventFilter(self)
-        slider_layout.addWidget(self.color_slider, 0, 0, 1, 3)
-
-        self.min_label = QLabel('Black')
-        self.min_label.setFont(self.gui.list_font)
-        slider_layout.addWidget(self.min_label, 1, 0, Qt.AlignmentFlag.AlignLeft)
-
-        self.color_label = QLabel()
-        color_pixmap = QPixmap(20, 20)
-        color_pixmap.fill(QColor(self.color_slider.value(), self.color_slider.value(), self.color_slider.value()))
-        self.color_label.setPixmap(color_pixmap)
-        self.color_slider.sliderMoved.connect(lambda value: self.change_sample(value))
-        slider_layout.addWidget(self.color_label, 1, 1, Qt.AlignmentFlag.AlignCenter)
-
-        self.max_label = QLabel('White')
-        self.max_label.setFont(self.gui.list_font)
-        slider_layout.addWidget(self.max_label, 1, 2, Qt.AlignmentFlag.AlignRight)
-
-    def change_sample(self, value):
-        new_pixmap = QPixmap(20, 20)
-        new_pixmap.fill(QColor(value, value, value))
-        self.color_label.setPixmap(new_pixmap)
-
-    def eventFilter(self, obj, evt):
-        if obj == self.color_slider and evt.type() == QEvent.Type.Wheel:
-            return True
-        elif obj == self.color_slider and evt.type() == QEvent.Type.MouseButtonRelease:
-            parent = self.parent()
-            while parent.parent():
-                if hasattr(parent, 'mouse_release_signal'):
-                    parent.mouse_release_signal.emit(self.color_slider.value())
-                    break
-                else:
-                    parent = parent.parent()
-            return super().eventFilter(obj, evt)
-        else:
-            return super().eventFilter(obj, evt)
-
-
-class OffsetSlider(QWidget):
-    """
-    Creates a widget containing a QSlider and Label which lets the user set the distance of the display's shadow offset
-    :param gui.GUI gui: The current instance of GUI
-    """
+class ClickableColorSwatch(QLabel):
+    color_changed = pyqtSignal()
     def __init__(self, gui):
         super().__init__()
         self.gui = gui
-        layout = QVBoxLayout()
-        layout.setContentsMargins(0, 0, 0, 0)
-        layout.setSpacing(0)
-        self.setLayout(layout)
 
-        self.offset_title = QLabel('Shadow Offset:')
-        self.offset_title.setFont(self.gui.list_font)
-        layout.addWidget(self.offset_title)
-
-        slider_widget = QWidget()
-        slider_widget.setFixedWidth(300)
-        slider_layout = QGridLayout()
-        slider_layout.setContentsMargins(0, 0, 0, 0)
-        slider_layout.setVerticalSpacing(0)
-        slider_widget.setLayout(slider_layout)
-        layout.addWidget(slider_widget)
-
-        self.offset_slider = CustomSlider()
-        self.offset_slider.setOrientation(Qt.Orientation.Horizontal)
-        self.offset_slider.setFont(self.gui.list_font)
-        self.offset_slider.setRange(0, 15)
-        self.offset_slider.setValue(self.gui.shadow_offset)
-        self.offset_slider.installEventFilter(self)
-        slider_layout.addWidget(self.offset_slider, 0, 0, 1, 3)
-
-        self.min_label = QLabel(str(self.offset_slider.minimum()) + 'px')
-        self.min_label.setFont(self.gui.list_font)
-        slider_layout.addWidget(self.min_label, 1, 0, Qt.AlignmentFlag.AlignLeft)
-
-        self.current_label = QLabel(str(self.offset_slider.value()) + 'px')
-        self.current_label.setFont(self.gui.list_title_font)
-        slider_layout.addWidget(self.current_label, 1, 1, Qt.AlignmentFlag.AlignCenter)
-        self.offset_slider.sliderMoved.connect(lambda value: self.current_label.setText(str(value) + 'px'))
-
-        self.max_label = QLabel(str(self.offset_slider.maximum()) + 'px')
-        self.max_label.setFont(self.gui.list_font)
-        slider_layout.addWidget(self.max_label, 1, 2, Qt.AlignmentFlag.AlignRight)
-
-    def eventFilter(self, obj, evt):
-        if obj == self.offset_slider and evt.type() == QEvent.Type.Wheel:
-            return True
-        elif obj == self.offset_slider and evt.type() == QEvent.Type.MouseButtonRelease:
-            parent = self.parent()
-            while parent.parent():
-                if hasattr(parent, 'mouse_release_signal'):
-                    parent.mouse_release_signal.emit(self.offset_slider.value())
-                    break
-                else:
-                    parent = parent.parent()
-            return super().eventFilter(obj, evt)
+    def make_color_swatch_pixmap(self, rgb_color):
+        if 'rgba' in rgb_color:
+            bg_color = rgb_color.replace('rgba(', '').replace(')', '')
         else:
-            return super().eventFilter(obj, evt)
+            bg_color = rgb_color.replace('rgb(', '').replace(')', '')
+        bg_color_split = bg_color.split(', ')
+
+        if len(bg_color_split) == 4:
+            brush = QBrush(
+                QColor(int(bg_color_split[0]), int(bg_color_split[1]), int(bg_color_split[2]), int(bg_color_split[3])))
+        else:
+            brush = QBrush(
+                QColor(int(bg_color_split[0]), int(bg_color_split[1]), int(bg_color_split[2])))
+        brush.setStyle(Qt.BrushStyle.SolidPattern)
+        pen = QPen(Qt.GlobalColor.black)
+        pen.setWidth(2)
+
+        pixmap = QPixmap(48, 48)
+        painter = QPainter(pixmap)
+        painter.setBrush(brush)
+        painter.setPen(pen)
+        painter.setPen(Qt.GlobalColor.black)
+
+        painter.begin(pixmap)
+        painter.fillRect(0, 0, 48, 48, brush)
+        painter.drawRect(QRect(0, 0, 48, 48))
+        painter.end()
+
+        self.setPixmap(pixmap)
+        self.repaint()
+
+    def mouseReleaseEvent(self, evt):
+        super().mouseReleaseEvent(evt)
+        image = self.pixmap().toImage()
+        current_color = image.pixelColor(10, 10)
+        chosen_color = QColorDialog.getColor(current_color, self.gui.main_window, 'Countdown Background Color')
+        rgb_color = f'rgba({chosen_color.red()}, {chosen_color.green()}, {chosen_color.blue()}, {chosen_color.alpha()})'
+        self.make_color_swatch_pixmap(rgb_color)
+        self.color_changed.emit()
+
+
+class CountdownWidget(QWidget):
+    update_label_signal = pyqtSignal(str)
+    show_self_signal = pyqtSignal()
+    hide_self_signal = pyqtSignal()
+
+    def __init__(self, gui, font, position, bg, fg):
+        super().__init__()
+        self.gui = gui
+
+        self.update_label_signal.connect(self.update_label)
+        self.show_self_signal.connect(self.show_self)
+        self.hide_self_signal.connect(self.hide_self)
+
+        #self.setParent(self.gui.display_widget)
+        self.setWindowFlag(Qt.WindowType.ToolTip)
+        self.setWindowFlag(Qt.WindowType.FramelessWindowHint)
+
+        self.setStyleSheet('background-color: ' + bg)
+        self.setAttribute(Qt.WA_StyledBackground, True)
+        layout = QGridLayout(self)
+
+        self.label = QLabel()
+        self.label.setFont(font)
+        self.label.setStyleSheet('color: ' + fg)
+        layout.addWidget(self.label, 0, 0, Qt.AlignmentFlag.AlignCenter)
+
+        font_metrics = QFontMetrics(font)
+        font_height = font_metrics.height()
+        height = font_height + 40
+
+        if position == 'top_full':
+            x = gui.display_widget.x()
+            y = gui.display_widget.y()
+            width = gui.display_widget.width()
+        elif position == 'bottom_full':
+            x = gui.display_widget.x()
+            y = gui.display_widget.y() + gui.display_widget.height() - height
+            width = gui.display_widget.width()
+
+        self.setGeometry(QRect(x, y, width, height))
+
+    def update_label(self, text):
+        self.label.setText(text)
+
+    def show_self(self):
+        self.show()
+        #self.raise_()
+        self.gui.main.app.processEvents()
+
+    def hide_self(self):
+        self.hide()
+        self.gui.main.app.processEvents()
 
 
 class CustomMainWindow(QMainWindow):
@@ -441,6 +218,20 @@ class CustomScrollArea(QScrollArea):
         self.widget().setFixedWidth(self.width())
 
 
+class CustomSlider(QSlider):
+    def __init__(self):
+        super().__init__()
+        self.mouse_pressed = False
+
+    def mousePressEvent(self, evt):
+        self.mouse_pressed = True
+        super().mousePressEvent(evt)
+
+    def mouseReleaseEvent(self, evt):
+        self.mouse_pressed = False
+        super().mouseReleaseEvent(evt)
+
+
 class DisplayWidget(QWidget):
     """
     Provides a custom QWidget to be used as the display widget
@@ -525,39 +316,96 @@ class DisplayWidget(QWidget):
             self.background_label.setPixmap(pixmap)
 
 
-class LyricDisplayWidget(QWidget):
+class FontFaceListWidget(QListWidget):
     """
-    Provide a standardized QWidget to be used for showing lyrics on the display and sample widgets.py
+    Creates a custom QListWidget that displays all fonts on the system in their own style.
+    :param gui.GUI gui: The current instance of GUI
     """
-    def __init__(
-            self,
-            gui,
-            for_sample=False,
-            use_outline=True,
-            outline_color=QColor(0, 0, 0),
-            outline_width=8,
-            fill_color=QColor(255, 255, 255),
-            use_shadow=True,
-            shadow_color=QColor(0, 0, 0),
-            shadow_offset=5,
-            use_shade=False,
-            shade_color=0,
-            shade_opacity=75):
+    def __init__(self, gui):
         """
-        Provide a standardized QWidget to be used for showing lyrics on the display and sample widgets.py
         :param gui.GUI gui: The current instance of GUI
-        :param bool for_sample: Whether this widget is intended for the sample widget or not
-        :param bool use_outline: Whether the font is to be outlined
-        :param int outline_color: The shade of the font outline (QColor(x, x, x))
-        :param int outline_width: The width, in px, of the font outline
-        :param fill_color: The fill color of the font
-        :param use_shadow: Whether the font is to be shadowed
-        :param shadow_color: The shade of the font shadow (QColor(x, x, x))
-        :param shadow_offset: The offset, in px, of the shadow
         """
         super().__init__()
         self.gui = gui
-        self.for_sample = for_sample
+        self.setObjectName('FontFaceListWidget')
+        self.setMinimumHeight(60)
+        self.blockSignals(True)
+        self.populate_widget()
+        self.blockSignals(False)
+
+    def populate_widget(self):
+        try:
+            families = QFontDatabase().families()
+            for font in families:
+                if self.gui.main.initial_startup:
+                    self.gui.main.update_status_signal.emit('Processing Fonts', 'status')
+                    self.gui.main.update_status_signal.emit(font, 'info')
+                list_label = QLabel(font)
+                list_label.setFont(QFont(font, 12))
+                item = QListWidgetItem()
+                item.setData(20, font)
+                self.addItem(item)
+                self.setItemWidget(item, list_label)
+
+            if self.gui.main.initial_startup:
+                self.gui.main.update_status_signal.emit('', 'info')
+        except Exception:
+            self.gui.main.error_log()
+
+
+class FontFaceComboBox(QComboBox):
+    """
+    Creates a custom QComboBox that displays all fonts on the system in their own style.
+    :param gui.GUI gui: The current instance of GUI
+    """
+    current_font = None
+
+    def __init__(self, gui):
+        """
+        :param gui.GUI gui: The current instance of GUI
+        """
+        super().__init__()
+        self.setObjectName('FontFaceComboBox')
+        self.gui = gui
+        self.setEditable(True)
+        self.setSizeAdjustPolicy(QComboBox.AdjustToMinimumContentsLengthWithIcon)
+        self.populate_widget()
+
+    def populate_widget(self):
+        try:
+            for i in range(len(self.gui.font_pixmaps)):
+                if i == len(self.gui.font_pixmaps) - 1:
+                    self.setIconSize(QSize(self.gui.font_pixmaps[i][0], self.gui.font_pixmaps[i][1]))
+                    self.setMinimumWidth(self.gui.font_pixmaps[i][0])
+                else:
+                    self.addItem(QIcon(self.gui.font_pixmaps[i][1]), self.gui.font_pixmaps[i][0])
+        except Exception:
+            self.gui.main.error_log()
+
+        for i in range(self.count()):
+            if self.itemText(i) == self.current_font:
+                self.setCurrentIndex(i)
+                break
+
+    def wheelEvent(self, evt):
+        evt.ignore()
+
+
+class FontSample(QLabel):
+    text = ''
+    def __init__(self, settings_widget,
+                 use_outline=True,
+                 outline_color=QColor(0, 0, 0),
+                 outline_width=8,
+                 fill_color=QColor(255, 255, 255),
+                 use_shadow=True,
+                 shadow_color=QColor(0, 0, 0),
+                 shadow_offset=5,
+                 use_shade=False,
+                 shade_color=0,
+                 shade_opacity=50):
+        super().__init__()
+        self.settings_widget = settings_widget
         self.use_outline = use_outline
         self.outline_color = outline_color
         self.outline_width = outline_width
@@ -569,266 +417,119 @@ class LyricDisplayWidget(QWidget):
         self.shade_color = shade_color
         self.shade_opacity = shade_opacity
 
-        self.text = ''
-        self.total_height = 0
+        self.sample_background = None
 
-        margins = QMargins(0, 0, 0, 0)
-        self.setContentsMargins(margins)
-
-        layout = QGridLayout()
-        layout.setRowStretch(0, 20)
-        layout.setRowStretch(1, 1)
-        layout.setContentsMargins(margins)
-        self.setLayout(layout)
-        self.setObjectName('lyric_display_widget')
-        self.setAutoFillBackground(False)
-
-        self.footer_label = QLabel()
-        self.footer_label.setContentsMargins(margins)
-        self.footer_label.setWordWrap(True)
-        self.footer_label.setObjectName('footer_label')
-        layout.addWidget(self.footer_label, 1, 0)
-
-    def setText(self, text):
-        """
-        Convenience method to set the text variable
-        :param str text: Text to be shown
-        """
-        self.text = text
-
-    def set_geometry(self):
-        """
-        Sets the geometry to match its parent widget
-        """
-        if self.for_sample:
-            self.setParent(self.gui.sample_widget)
-            self.setGeometry(self.gui.sample_widget.geometry())
-            self.move(self.gui.sample_widget.x(), self.gui.sample_widget.y())
+        self.container = self.settings_widget.findChild(QWidget, 'font_sample_container')
+        self.widget = self.settings_widget.findChild(QWidget, 'font_sample_widget')
+        self.background_label = self.settings_widget.findChild(QLabel, 'font_sample_background_label')
 
     def paintEvent(self, evt):
-        """
-        Overrides paintEvent to custom paint the text onto the widget
-        :param QPaintEvent evt: paintEvent
-        """
-        palette = self.footer_label.palette()
-        palette.setColor(QPalette.ColorRole.WindowText, self.fill_color)
-        self.footer_label.setPalette(palette)
-        self.paint_text()
+        self.paint_font()
+        super().paintEvent(evt)
 
-    def paint_text(self):
-        """
-        Method to paint the text onto the widget, using the prescribed outline and shadow values
-        """
-        self.total_height = 0
-        self.text = re.sub('<p.*?>', '', self.text)
-        self.text = re.sub('</p>', '', self.text)
-        self.text = re.sub('\n', '<br />', self.text)
-        self.text = re.sub('<br/>', '<br />', self.text)
-
-        BOLD = 0
-        ITALIC = 1
-        UNDERLINE = 2
-
-        """text_lines = self.text.split('<br />')
-        line_height = self.fontMetrics().boundingRect('Way').height()
-        lines = []
-        for this_line in text_lines:
-            this_line = this_line.strip()
-
-            # split the lines according to their drawn length
-            if self.fontMetrics().boundingRect(this_line).width() < self.gui.display_widget.width() - 20:
-                lines.append(this_line)
-            else:
-                this_line_split = this_line.split(' ')
-                current_line = ''
-                for word in this_line_split:
-                    current_line = (current_line + ' ' + word).strip()
-                    print(current_line)
-                    if self.fontMetrics().boundingRect(current_line).width() > self.gui.display_widget.width() - 20:
-                        lines.append(' '.join(current_line.split(' ')[:-1]))
-                        current_line = word
-                lines.append(current_line)"""
-
-        font = self.font()
-        font_size = font.pointSize() + 2
-        painter_paths = []
-        longest_line = 0
-        self.footer_label.adjustSize()
-
-        # build paths for each line, creating a new path whenever the line becomes too long
-        footer_height = self.footer_label.height()
-        if self.footer_label.isHidden() or len(self.footer_label.text().strip()) == 0:
-            footer_height = 0
-        usable_rect = QRect(0, 0, self.gui.display_widget.width(), self.gui.display_widget.height() - footer_height - 40)
-        self.total_height = -1
-        while self.total_height == -1 or self.total_height > usable_rect.height():
-            longest_line = 0
-            painter_paths = []
-            word_path = QPainterPath()
-            path_index = -1
-
-            font_size -= 2
-            font = QFont(font.family(), font_size)
-            self.setFont(font)
-            line_height = self.fontMetrics().boundingRect('Way').height()
-            space_width = self.fontMetrics().boundingRect('w w').width() - self.fontMetrics().boundingRect('ww').width()
-
-            lines = self.text.split('<br />')
-            for i in range(len(lines)):
-                #if len(re.sub('<.*?>', '', lines[i]).strip()) > 0:
-                x = 0
-                y = 0
-                line_words = lines[i].split(' ')
-                if len(line_words) == 0:
-                    line_words = [' ']
-                painter_paths.append(QPainterPath())
-                path_index += 1
-                for word in line_words:
-                    #if len(re.sub('<.*?>', '', word).strip()) > 0:
-                    word_path.clear()
-                    if '<b>' in word:
-                        font.setWeight(1000)
-                    if '<i>' in word:
-                        font.setItalic(True)
-                    if '<u>' in word:
-                        font.setUnderline(True)
-
-                    word_path.addText(QPointF(x, y), font, re.sub('<.*?>', '', word))
-                    if (painter_paths[path_index].boundingRect().width() + word_path.boundingRect().width()
-                            > self.gui.display_widget.width() - 40):
-                        painter_paths.append(QPainterPath())
-                        x = 0
-                        y = 0
-                        path_index += 1
-                    painter_paths[path_index].addText(QPointF(x, y), font, re.sub('<.*?>', '', word))
-                    x = painter_paths[path_index].boundingRect().width() + space_width
-
-                    if '</b>' in word:
-                        font.setWeight(QFont.Weight.Normal)
-                    if '</i>' in word:
-                        font.setItalic(False)
-                    if '</u>' in word:
-                        font.setUnderline(False)
-
-            # get the total size of the paths that will be drawn for creating the shading rectangle
-            self.total_height = 0
-            for path in painter_paths:
-                #if path.boundingRect().width() > 0:
-                self.total_height += line_height
-                if path.boundingRect().width() > longest_line:
-                    longest_line = path.boundingRect().width()
-
-            if self.for_sample:
-                break
-
-        # start the first path at the midpoint of the usable rect, minus half the total height of the paths, plus
-        # the font's ascent (to account for the path's y being the baseline of the text) plus a 20px margin at the top
-        path_y = (usable_rect.height() / 2) - (self.total_height / 2) + self.fontMetrics().ascent() + 20
-        starting_y = path_y
-        painter = QPainter(self)
+    def paint_font(self):
         brush = QBrush()
-        painter.setBrush(brush)
         pen = QPen()
-        painter.setPen(pen)
+
+        path = QPainterPath()
+        shadow_path = QPainterPath()
+        metrics = self.fontMetrics()
+
+        y = metrics.ascent() - metrics.descent() + 20
+        point = QPointF(20, y)
+        shadow_point = QPointF(point.x() + self.shadow_offset, point.y() + self.shadow_offset)
+
+        if self.use_shadow:
+            shadow_path.addText(shadow_point, self.font(), self.text)
+        path.addText(point, self.font(), self.text)
+        path_rect = path.boundingRect()
+
+        image_rect = QRectF(0, 0, path_rect.width() + 40, path_rect.height() + 40)
+        image = QPixmap(int(image_rect.width()), int(image_rect.height()))
+
+        painter = QPainter(image)
         painter.setRenderHint(QPainter.RenderHint.Antialiasing)
+        painter.begin(image)
+
+        background_image = self.make_sample_background(image_rect)
+        painter.drawImage(QPoint(0, 0), background_image)
 
         opacity = self.shade_opacity
         if not self.use_shade:
             opacity = 0
-        shade_rect = QRectF(
-            int((self.gui.display_widget.width() / 2) - (longest_line / 2)) - 20,
-            starting_y - self.fontMetrics().ascent() - 20,
-            longest_line + 40,
-            self.total_height + 40
-        )
-        painter.fillRect(shade_rect, QColor(self.shade_color, self.shade_color, self.shade_color, opacity))
+        brush.setColor(QColor(self.shade_color, self.shade_color, self.shade_color, opacity))
+        brush.setStyle(Qt.BrushStyle.SolidPattern)
+        shade_rect = QRectF(10, 10, path_rect.width() + 20, path_rect.height() + 20)
+        rect_item = QGraphicsRectItem(shade_rect)
+        rect_item.setBrush(brush)
+        painter.fillRect(shade_rect, brush)
 
-        for path in painter_paths:
-            #if path.boundingRect().width() > 0:
-            path_x = (self.gui.display_widget.width() / 2) - (path.boundingRect().width() / 2)
-            path.translate(path_x, path_y)
+        if self.use_shadow:
+            brush.setColor(self.shadow_color)
+            pen.setWidth(0)
+            painter.fillPath(shadow_path, brush)
 
-            if self.use_shadow:
-                path.translate(self.shadow_offset, self.shadow_offset)
-                shadow_brush = QBrush()
-                shadow_brush.setColor(self.shadow_color)
-                shadow_brush.setStyle(Qt.BrushStyle.SolidPattern)
-                painter.fillPath(path, shadow_brush)
-                path.translate(-self.shadow_offset, -self.shadow_offset)
+        brush.setColor(self.fill_color)
+        pen.setColor(self.outline_color)
+        pen.setWidth(self.outline_width)
+        painter.fillPath(path, brush)
 
-            brush.setColor(self.fill_color)
-            brush.setStyle(Qt.BrushStyle.SolidPattern)
-            pen.setColor(self.outline_color)
-            pen.setWidth(self.outline_width)
+        if self.use_outline:
             painter.setPen(pen)
+            painter.drawPath(path)
+        painter.end()
 
-            painter.fillPath(path, brush)
-            if self.use_outline:
-                painter.strokePath(path, pen)
+        self.setPixmap(image)
 
-            path_y += line_height
+    def make_sample_background(self, rect):
+        slide_type = self.settings_widget.slide_type
 
+        if self.settings_widget.applies_to_global:
+            sample_background = QImage(
+                self.settings_widget.gui.main.background_dir + '/'
+                + self.settings_widget.gui.main.settings[f'global_{slide_type}_background'])
+        else:
+            background = self.settings_widget.parent().parent().findChild(QLineEdit, 'background_line_edit').text()
+            if 'rgb(' in background:
+                background = background.replace('rgb(', '')
+                background = background.replace(')', '')
+                background_split = background.split(', ')
+                sample_background = QImage(QSize(1920, 1080), QImage.Format.Format_RGB32)
+                sample_background.fill(
+                    QColor(int(background_split[0]), int(background_split[1]), int(background_split[2])))
+            elif 'Song' in background:
+                sample_background = QImage(
+                    self.settings_widget.gui.main.background_dir + '/'
+                    + self.settings_widget.gui.main.settings['global_song_background'])
+            elif 'Bible' in background:
+                sample_background = QImage(
+                    self.settings_widget.gui.main.background_dir + '/'
+                    + self.settings_widget.gui.main.settings['global_bible_background'])
+            else:
+                sample_background = QImage(self.settings_widget.gui.main.background_dir + '/' + background)
 
-class StandardItemWidget(QWidget):
-    """
-    Provides a standardized QWidget to be used as a QListWidget ItemWidget
-    """
-    def __init__(self, gui, title, subtitle=None, icon=None, wrap_subtitle=False):
-        super().__init__()
-        self.gui = gui
-        self.setObjectName('item_widget')
-        layout = QHBoxLayout(self)
+        ratio = sample_background.width() / rect.width()
+        # if there was no background yet chosen, ration will be 0
+        if ratio > 0:
+            sample_background = sample_background.scaled(
+                int(rect.width()),
+                int(sample_background.height() / ratio),
+                Qt.AspectRatioMode.IgnoreAspectRatio,
+                Qt.TransformationMode.SmoothTransformation
+            )
 
-        self.subtitle = None
-        self.icon = None
+            piece_rect = QRect(
+                0,
+                int(sample_background.height() / 2) - int(rect.height() / 2),
+                int(rect.width()),
+                int(rect.height())
+            )
+            sample_background = sample_background.copy(piece_rect)
+        else:
+            sample_background = QImage(QSize(int(rect.width()), int(rect.height())), QImage.Format_RGB32)
+            sample_background.fill(Qt.GlobalColor.black)
 
-        if icon:
-            self.icon = QLabel()
-            self.icon.setAutoFillBackground(False)
-            self.icon.setPixmap(icon)
-            self.icon.adjustSize()
-            layout.addWidget(self.icon)
-
-        text_container = QWidget()
-        text_layout = QVBoxLayout(text_container)
-        layout.addWidget(text_container)
-
-        self.title = QLabel(title)
-        self.title.setAutoFillBackground(False)
-        self.title.setObjectName('lyric_item_widget_title')
-        self.title.setFont(self.gui.list_title_font)
-        self.title.adjustSize()
-        text_layout.addWidget(self.title)
-
-        if subtitle:
-            subtitle = re.sub('<br.*?>', '\n', subtitle)
-            subtitle = re.sub('<.*?>', '', subtitle)
-
-            self.subtitle = QLabel(subtitle)
-            self.subtitle.setAutoFillBackground(False)
-            self.title.setObjectName('lyric_item_widget_text')
-            if wrap_subtitle:
-                self.subtitle.setWordWrap(True)
-            self.subtitle.setFont(self.gui.list_font)
-            self.subtitle.adjustSize()
-            text_layout.addWidget(self.subtitle)
-
-        if not wrap_subtitle:
-            layout.addStretch()
-
-        self.adjustSize()
-
-
-class AutoSelectLineEdit(QLineEdit):
-    """
-    Implements QLineEdit to add the ability to select all text when this line edit receives focus.
-    """
-    def __init__(self):
-        super().__init__()
-
-    def focusInEvent(self, evt):
-        super().focusInEvent(evt)
-        QTimer.singleShot(0, self.selectAll)
+        return sample_background
 
 
 class FontWidget(QWidget):
@@ -1238,6 +939,376 @@ class FontWidget(QWidget):
         super().hideEvent(evt)
 
 
+class ImageCombobox(QComboBox):
+    """
+    Creates a custom QComboBox that displays a thumbnail of an image to be used.
+    """
+    def __init__(self, gui, type, suppress_autosave=False):
+        """
+        :param gui.GUI gui: The current instance of GUI
+        :param str type: Whether this is creating a combobox of 'logo', 'song', or 'bible' images
+        """
+        super().__init__()
+        self.gui = gui
+        self.type = type
+        self.table = None
+        self.suppress_autosave = suppress_autosave
+        self.setView(QListView())
+        self.setObjectName(type)
+
+        self.setIconSize(QSize(96, 54))
+        self.setMaximumWidth(240)
+        self.setFont(self.gui.standard_font)
+
+        self.currentIndexChanged.connect(self.index_changed)
+
+        if type == 'edit_background':
+            self.removeItem(0)
+        else:
+            self.currentIndexChanged.connect(self.gui.tool_bar.change_background)
+        self.refresh()
+
+    def index_changed(self):
+        file_name = self.itemData(self.currentIndex(), Qt.ItemDataRole.UserRole)
+        if not file_name:
+            return
+        if self.type == 'logo':
+            self.gui.main.settings['logo_image'] = file_name
+        elif self.type == 'song':
+            self.gui.main.settings['global_song_background'] = file_name
+            self.gui.global_song_background_pixmap = QPixmap(self.gui.main.image_dir + '/' + file_name)
+        elif self.type == 'bible':
+            self.gui.main.settings['global_bible_background'] = file_name
+            self.gui.global_bible_background_pixmap = QPixmap(self.gui.main.image_dir + '/' + file_name)
+
+        for i in range(self.gui.oos_widget.oos_list_widget.count()):
+            item = self.gui.oos_widget.oos_list_widget.item(i)
+            item_data = item.data(Qt.ItemDataRole.UserRole)
+            pixmap = None
+
+            if item_data['type'] == 'song' or item_data['type'] == 'bible' or item_data['type'] == 'custom':
+                if not item_data['override_global'] or item_data['override_global'] == 'False':
+                    if item_data['type'] == self.type:
+                        pixmap = self.itemIcon(self.currentIndex()).pixmap(QSize(50, 27))
+                    elif item_data['type'] == 'custom' and self.type == 'bible':
+                        pixmap = self.itemIcon(self.currentIndex()).pixmap(QSize(50, 27))
+                else:
+                    if item_data['background'] == 'global_song' and self.type == 'song':
+                        pixmap = self.itemIcon(self.currentIndex()).pixmap(QSize(50, 27))
+                    elif item_data['background'] == 'global_bible' and self.type == 'bible':
+                        pixmap = self.itemIcon(self.currentIndex()).pixmap(QSize(50, 27))
+
+            if pixmap:
+                widget = self.gui.oos_widget.oos_list_widget.itemWidget(item)
+                widget.icon.setPixmap(pixmap)
+                widget.adjustSize()
+                item.setSizeHint(widget.sizeHint())
+
+        if not self.suppress_autosave:
+            self.gui.main.save_settings()
+
+    def refresh(self):
+        """
+        Method to refresh the combo box after changes to the image indices
+        """
+        self.blockSignals(True)
+        self.clear()
+
+        if self.type == 'logo':
+            self.addItem('Choose Logo Image', userData='choose_logo')
+            self.addItem('Import a Logo Image', userData='import_logo')
+            self.table = 'imageThumbnails'
+        elif self.type == 'edit':
+            self.addItem('Choose Custom Background', userData='choose_global')
+            self.table = 'backgroundThumbnails'
+        elif self.type == 'delete_background':
+            self.addItem('Choose Background to Remove')
+            self.table = 'backgroundThumbnails'
+        elif self.type == 'delete_image':
+            self.addItem('Choose Image Item to Remove')
+            self.table = 'imageThumbnails'
+        else:
+            self.addItem('Choose Global ' + self.type + ' Background', userData='choose_global')
+            self.addItem('Import a Background Image', userData='import_global')
+            self.table = 'backgroundThumbnails'
+        connection = None
+
+        try:
+            image_list = []
+            connection = sqlite3.connect(self.gui.main.database)
+            cursor = connection.cursor()
+            thumbnails = cursor.execute(
+                'SELECT * FROM ' + self.table + ' ORDER BY fileName COLLATE NOCASE ASC;').fetchall()
+            for record in thumbnails:
+                if self.gui.main.initial_startup:
+                    self.gui.main.update_status_signal.emit('Loading Thumbnails', 'status')
+                    self.gui.main.update_status_signal.emit(record[0], 'info')
+                pixmap = QPixmap()
+                pixmap.loadFromData(record[1], 'JPG')
+                icon = QIcon(pixmap)
+                self.addItem(icon, record[0].split('.')[0], userData=record[0])
+                image_list.append([icon, record[0].split('.')[0], record[0]])
+            connection.close()
+
+            if self.gui.main.initial_startup:
+                self.gui.main.update_status_signal.emit('', 'info')
+
+            self.blockSignals(False)
+        except Exception:
+            self.gui.main.error_log()
+            if connection:
+                connection.close()
+            self.blockSignals(False)
+
+    def wheelEvent(self, evt):
+        # prevent wheel scrolling, which is undesirable in the settings layout
+        evt.ignore()
+
+
+class LyricDisplayWidget(QWidget):
+    """
+    Provide a standardized QWidget to be used for showing lyrics on the display and sample widgets.py
+    """
+    def __init__(
+            self,
+            gui,
+            for_sample=False,
+            use_outline=True,
+            outline_color=QColor(0, 0, 0),
+            outline_width=8,
+            fill_color=QColor(255, 255, 255),
+            use_shadow=True,
+            shadow_color=QColor(0, 0, 0),
+            shadow_offset=5,
+            use_shade=False,
+            shade_color=0,
+            shade_opacity=75):
+        """
+        Provide a standardized QWidget to be used for showing lyrics on the display and sample widgets.py
+        :param gui.GUI gui: The current instance of GUI
+        :param bool for_sample: Whether this widget is intended for the sample widget or not
+        :param bool use_outline: Whether the font is to be outlined
+        :param int outline_color: The shade of the font outline (QColor(x, x, x))
+        :param int outline_width: The width, in px, of the font outline
+        :param fill_color: The fill color of the font
+        :param use_shadow: Whether the font is to be shadowed
+        :param shadow_color: The shade of the font shadow (QColor(x, x, x))
+        :param shadow_offset: The offset, in px, of the shadow
+        """
+        super().__init__()
+        self.gui = gui
+        self.for_sample = for_sample
+        self.use_outline = use_outline
+        self.outline_color = outline_color
+        self.outline_width = outline_width
+        self.fill_color = fill_color
+        self.use_shadow = use_shadow
+        self.shadow_color = shadow_color
+        self.shadow_offset = shadow_offset
+        self.use_shade = use_shade
+        self.shade_color = shade_color
+        self.shade_opacity = shade_opacity
+
+        self.text = ''
+        self.total_height = 0
+
+        margins = QMargins(0, 0, 0, 0)
+        self.setContentsMargins(margins)
+
+        layout = QGridLayout()
+        layout.setRowStretch(0, 20)
+        layout.setRowStretch(1, 1)
+        layout.setContentsMargins(margins)
+        self.setLayout(layout)
+        self.setObjectName('lyric_display_widget')
+        self.setAutoFillBackground(False)
+
+        self.footer_label = QLabel()
+        self.footer_label.setContentsMargins(margins)
+        self.footer_label.setWordWrap(True)
+        self.footer_label.setObjectName('footer_label')
+        layout.addWidget(self.footer_label, 1, 0)
+
+    def setText(self, text):
+        """
+        Convenience method to set the text variable
+        :param str text: Text to be shown
+        """
+        self.text = text
+
+    def set_geometry(self):
+        """
+        Sets the geometry to match its parent widget
+        """
+        if self.for_sample:
+            self.setParent(self.gui.sample_widget)
+            self.setGeometry(self.gui.sample_widget.geometry())
+            self.move(self.gui.sample_widget.x(), self.gui.sample_widget.y())
+
+    def paintEvent(self, evt):
+        """
+        Overrides paintEvent to custom paint the text onto the widget
+        :param QPaintEvent evt: paintEvent
+        """
+        palette = self.footer_label.palette()
+        palette.setColor(QPalette.ColorRole.WindowText, self.fill_color)
+        self.footer_label.setPalette(palette)
+        self.paint_text()
+
+    def paint_text(self):
+        """
+        Method to paint the text onto the widget, using the prescribed outline and shadow values
+        """
+        self.total_height = 0
+        self.text = re.sub('<p.*?>', '', self.text)
+        self.text = re.sub('</p>', '', self.text)
+        self.text = re.sub('\n', '<br />', self.text)
+        self.text = re.sub('<br/>', '<br />', self.text)
+
+        BOLD = 0
+        ITALIC = 1
+        UNDERLINE = 2
+
+        """text_lines = self.text.split('<br />')
+        line_height = self.fontMetrics().boundingRect('Way').height()
+        lines = []
+        for this_line in text_lines:
+            this_line = this_line.strip()
+
+            # split the lines according to their drawn length
+            if self.fontMetrics().boundingRect(this_line).width() < self.gui.display_widget.width() - 20:
+                lines.append(this_line)
+            else:
+                this_line_split = this_line.split(' ')
+                current_line = ''
+                for word in this_line_split:
+                    current_line = (current_line + ' ' + word).strip()
+                    print(current_line)
+                    if self.fontMetrics().boundingRect(current_line).width() > self.gui.display_widget.width() - 20:
+                        lines.append(' '.join(current_line.split(' ')[:-1]))
+                        current_line = word
+                lines.append(current_line)"""
+
+        font = self.font()
+        font_size = font.pointSize() + 2
+        painter_paths = []
+        longest_line = 0
+        self.footer_label.adjustSize()
+
+        # build paths for each line, creating a new path whenever the line becomes too long
+        footer_height = self.footer_label.height()
+        if self.footer_label.isHidden() or len(self.footer_label.text().strip()) == 0:
+            footer_height = 0
+        usable_rect = QRect(0, 0, self.gui.display_widget.width(), self.gui.display_widget.height() - footer_height - 40)
+        self.total_height = -1
+        while self.total_height == -1 or self.total_height > usable_rect.height():
+            longest_line = 0
+            painter_paths = []
+            word_path = QPainterPath()
+            path_index = -1
+
+            font_size -= 2
+            font = QFont(font.family(), font_size)
+            self.setFont(font)
+            line_height = self.fontMetrics().boundingRect('Way').height()
+            space_width = self.fontMetrics().boundingRect('w w').width() - self.fontMetrics().boundingRect('ww').width()
+
+            lines = self.text.split('<br />')
+            for i in range(len(lines)):
+                #if len(re.sub('<.*?>', '', lines[i]).strip()) > 0:
+                x = 0
+                y = 0
+                line_words = lines[i].split(' ')
+                if len(line_words) == 0:
+                    line_words = [' ']
+                painter_paths.append(QPainterPath())
+                path_index += 1
+                for word in line_words:
+                    #if len(re.sub('<.*?>', '', word).strip()) > 0:
+                    word_path.clear()
+                    if '<b>' in word:
+                        font.setWeight(1000)
+                    if '<i>' in word:
+                        font.setItalic(True)
+                    if '<u>' in word:
+                        font.setUnderline(True)
+
+                    word_path.addText(QPointF(x, y), font, re.sub('<.*?>', '', word))
+                    if (painter_paths[path_index].boundingRect().width() + word_path.boundingRect().width()
+                            > self.gui.display_widget.width() - 40):
+                        painter_paths.append(QPainterPath())
+                        x = 0
+                        y = 0
+                        path_index += 1
+                    painter_paths[path_index].addText(QPointF(x, y), font, re.sub('<.*?>', '', word))
+                    x = painter_paths[path_index].boundingRect().width() + space_width
+
+                    if '</b>' in word:
+                        font.setWeight(QFont.Weight.Normal)
+                    if '</i>' in word:
+                        font.setItalic(False)
+                    if '</u>' in word:
+                        font.setUnderline(False)
+
+            # get the total size of the paths that will be drawn for creating the shading rectangle
+            self.total_height = 0
+            for path in painter_paths:
+                #if path.boundingRect().width() > 0:
+                self.total_height += line_height
+                if path.boundingRect().width() > longest_line:
+                    longest_line = path.boundingRect().width()
+
+            if self.for_sample:
+                break
+
+        # start the first path at the midpoint of the usable rect, minus half the total height of the paths, plus
+        # the font's ascent (to account for the path's y being the baseline of the text) plus a 20px margin at the top
+        path_y = (usable_rect.height() / 2) - (self.total_height / 2) + self.fontMetrics().ascent() + 20
+        starting_y = path_y
+        painter = QPainter(self)
+        brush = QBrush()
+        painter.setBrush(brush)
+        pen = QPen()
+        painter.setPen(pen)
+        painter.setRenderHint(QPainter.RenderHint.Antialiasing)
+
+        opacity = self.shade_opacity
+        if not self.use_shade:
+            opacity = 0
+        shade_rect = QRectF(
+            int((self.gui.display_widget.width() / 2) - (longest_line / 2)) - 20,
+            starting_y - self.fontMetrics().ascent() - 20,
+            longest_line + 40,
+            self.total_height + 40
+        )
+        painter.fillRect(shade_rect, QColor(self.shade_color, self.shade_color, self.shade_color, opacity))
+
+        for path in painter_paths:
+            #if path.boundingRect().width() > 0:
+            path_x = (self.gui.display_widget.width() / 2) - (path.boundingRect().width() / 2)
+            path.translate(path_x, path_y)
+
+            if self.use_shadow:
+                path.translate(self.shadow_offset, self.shadow_offset)
+                shadow_brush = QBrush()
+                shadow_brush.setColor(self.shadow_color)
+                shadow_brush.setStyle(Qt.BrushStyle.SolidPattern)
+                painter.fillPath(path, shadow_brush)
+                path.translate(-self.shadow_offset, -self.shadow_offset)
+
+            brush.setColor(self.fill_color)
+            brush.setStyle(Qt.BrushStyle.SolidPattern)
+            pen.setColor(self.outline_color)
+            pen.setWidth(self.outline_width)
+            painter.setPen(pen)
+
+            painter.fillPath(path, brush)
+            if self.use_outline:
+                painter.strokePath(path, pen)
+
+            path_y += line_height
+
+
 class NewFontWidget(QWidget):
     """
     Implements QWidget that contains all of the settings that can be applied to the display font
@@ -1628,260 +1699,66 @@ class NewFontWidget(QWidget):
         super().hideEvent(evt)
 
 
-class CustomSlider(QSlider):
-    def __init__(self):
-        super().__init__()
-        self.mouse_pressed = False
-
-    def mousePressEvent(self, evt):
-        self.mouse_pressed = True
-        super().mousePressEvent(evt)
-
-    def mouseReleaseEvent(self, evt):
-        self.mouse_pressed = False
-        super().mouseReleaseEvent(evt)
-
-
-class FontSample(QLabel):
-    text = ''
-    def __init__(self, settings_widget,
-                 use_outline=True,
-                 outline_color=QColor(0, 0, 0),
-                 outline_width=8,
-                 fill_color=QColor(255, 255, 255),
-                 use_shadow=True,
-                 shadow_color=QColor(0, 0, 0),
-                 shadow_offset=5,
-                 use_shade=False,
-                 shade_color=0,
-                 shade_opacity=50):
-        super().__init__()
-        self.settings_widget = settings_widget
-        self.use_outline = use_outline
-        self.outline_color = outline_color
-        self.outline_width = outline_width
-        self.fill_color = fill_color
-        self.use_shadow = use_shadow
-        self.shadow_color = shadow_color
-        self.shadow_offset = shadow_offset
-        self.use_shade = use_shade
-        self.shade_color = shade_color
-        self.shade_opacity = shade_opacity
-
-        self.sample_background = None
-
-        self.container = self.settings_widget.findChild(QWidget, 'font_sample_container')
-        self.widget = self.settings_widget.findChild(QWidget, 'font_sample_widget')
-        self.background_label = self.settings_widget.findChild(QLabel, 'font_sample_background_label')
-
-    def paintEvent(self, evt):
-        self.paint_font()
-        super().paintEvent(evt)
-
-    def paint_font(self):
-        brush = QBrush()
-        pen = QPen()
-
-        path = QPainterPath()
-        shadow_path = QPainterPath()
-        metrics = self.fontMetrics()
-
-        y = metrics.ascent() - metrics.descent() + 20
-        point = QPointF(20, y)
-        shadow_point = QPointF(point.x() + self.shadow_offset, point.y() + self.shadow_offset)
-
-        if self.use_shadow:
-            shadow_path.addText(shadow_point, self.font(), self.text)
-        path.addText(point, self.font(), self.text)
-        path_rect = path.boundingRect()
-
-        image_rect = QRectF(0, 0, path_rect.width() + 40, path_rect.height() + 40)
-        image = QPixmap(int(image_rect.width()), int(image_rect.height()))
-
-        painter = QPainter(image)
-        painter.setRenderHint(QPainter.RenderHint.Antialiasing)
-        painter.begin(image)
-
-        background_image = self.make_sample_background(image_rect)
-        painter.drawImage(QPoint(0, 0), background_image)
-
-        opacity = self.shade_opacity
-        if not self.use_shade:
-            opacity = 0
-        brush.setColor(QColor(self.shade_color, self.shade_color, self.shade_color, opacity))
-        brush.setStyle(Qt.BrushStyle.SolidPattern)
-        shade_rect = QRectF(10, 10, path_rect.width() + 20, path_rect.height() + 20)
-        rect_item = QGraphicsRectItem(shade_rect)
-        rect_item.setBrush(brush)
-        painter.fillRect(shade_rect, brush)
-
-        if self.use_shadow:
-            brush.setColor(self.shadow_color)
-            pen.setWidth(0)
-            painter.fillPath(shadow_path, brush)
-
-        brush.setColor(self.fill_color)
-        pen.setColor(self.outline_color)
-        pen.setWidth(self.outline_width)
-        painter.fillPath(path, brush)
-
-        if self.use_outline:
-            painter.setPen(pen)
-            painter.drawPath(path)
-        painter.end()
-
-        self.setPixmap(image)
-
-    def make_sample_background(self, rect):
-        slide_type = self.settings_widget.slide_type
-
-        if self.settings_widget.applies_to_global:
-            sample_background = QImage(
-                self.settings_widget.gui.main.background_dir + '/'
-                + self.settings_widget.gui.main.settings[f'global_{slide_type}_background'])
-        else:
-            background = self.settings_widget.parent().parent().findChild(QLineEdit, 'background_line_edit').text()
-            if 'rgb(' in background:
-                background = background.replace('rgb(', '')
-                background = background.replace(')', '')
-                background_split = background.split(', ')
-                sample_background = QImage(QSize(1920, 1080), QImage.Format.Format_RGB32)
-                sample_background.fill(
-                    QColor(int(background_split[0]), int(background_split[1]), int(background_split[2])))
-            elif 'Song' in background:
-                sample_background = QImage(
-                    self.settings_widget.gui.main.background_dir + '/'
-                    + self.settings_widget.gui.main.settings['global_song_background'])
-            elif 'Bible' in background:
-                sample_background = QImage(
-                    self.settings_widget.gui.main.background_dir + '/'
-                    + self.settings_widget.gui.main.settings['global_bible_background'])
-            else:
-                sample_background = QImage(self.settings_widget.gui.main.background_dir + '/' + background)
-
-        ratio = sample_background.width() / rect.width()
-        # if there was no background yet chosen, ration will be 0
-        if ratio > 0:
-            sample_background = sample_background.scaled(
-                int(rect.width()),
-                int(sample_background.height() / ratio),
-                Qt.AspectRatioMode.IgnoreAspectRatio,
-                Qt.TransformationMode.SmoothTransformation
-            )
-
-            piece_rect = QRect(
-                0,
-                int(sample_background.height() / 2) - int(rect.height() / 2),
-                int(rect.width()),
-                int(rect.height())
-            )
-            sample_background = sample_background.copy(piece_rect)
-        else:
-            sample_background = QImage(QSize(int(rect.width()), int(rect.height())), QImage.Format_RGB32)
-            sample_background.fill(Qt.GlobalColor.black)
-
-        return sample_background
-
-
-class CountdownWidget(QWidget):
-    update_label_signal = pyqtSignal(str)
-    show_self_signal = pyqtSignal()
-    hide_self_signal = pyqtSignal()
-
-    def __init__(self, gui, font, position, bg, fg):
-        super().__init__()
-        self.gui = gui
-
-        self.update_label_signal.connect(self.update_label)
-        self.show_self_signal.connect(self.show_self)
-        self.hide_self_signal.connect(self.hide_self)
-
-        #self.setParent(self.gui.display_widget)
-        self.setWindowFlag(Qt.WindowType.ToolTip)
-        self.setWindowFlag(Qt.WindowType.FramelessWindowHint)
-
-        self.setStyleSheet('background-color: ' + bg)
-        self.setAttribute(Qt.WA_StyledBackground, True)
-        layout = QGridLayout(self)
-
-        self.label = QLabel()
-        self.label.setFont(font)
-        self.label.setStyleSheet('color: ' + fg)
-        layout.addWidget(self.label, 0, 0, Qt.AlignmentFlag.AlignCenter)
-
-        font_metrics = QFontMetrics(font)
-        font_height = font_metrics.height()
-        height = font_height + 40
-
-        if position == 'top_full':
-            x = gui.display_widget.x()
-            y = gui.display_widget.y()
-            width = gui.display_widget.width()
-        elif position == 'bottom_full':
-            x = gui.display_widget.x()
-            y = gui.display_widget.y() + gui.display_widget.height() - height
-            width = gui.display_widget.width()
-
-        self.setGeometry(QRect(x, y, width, height))
-
-    def update_label(self, text):
-        self.label.setText(text)
-
-    def show_self(self):
-        self.show()
-        #self.raise_()
-        self.gui.main.app.processEvents()
-
-    def hide_self(self):
-        self.hide()
-        self.gui.main.app.processEvents()
-
-
-class ClickableColorSwatch(QLabel):
-    color_changed = pyqtSignal()
+class OffsetSlider(QWidget):
+    """
+    Creates a widget containing a QSlider and Label which lets the user set the distance of the display's shadow offset
+    :param gui.GUI gui: The current instance of GUI
+    """
     def __init__(self, gui):
         super().__init__()
         self.gui = gui
+        layout = QVBoxLayout()
+        layout.setContentsMargins(0, 0, 0, 0)
+        layout.setSpacing(0)
+        self.setLayout(layout)
 
-    def make_color_swatch_pixmap(self, rgb_color):
-        if 'rgba' in rgb_color:
-            bg_color = rgb_color.replace('rgba(', '').replace(')', '')
+        self.offset_title = QLabel('Shadow Offset:')
+        self.offset_title.setFont(self.gui.list_font)
+        layout.addWidget(self.offset_title)
+
+        slider_widget = QWidget()
+        slider_widget.setFixedWidth(300)
+        slider_layout = QGridLayout()
+        slider_layout.setContentsMargins(0, 0, 0, 0)
+        slider_layout.setVerticalSpacing(0)
+        slider_widget.setLayout(slider_layout)
+        layout.addWidget(slider_widget)
+
+        self.offset_slider = CustomSlider()
+        self.offset_slider.setOrientation(Qt.Orientation.Horizontal)
+        self.offset_slider.setFont(self.gui.list_font)
+        self.offset_slider.setRange(0, 15)
+        self.offset_slider.setValue(self.gui.shadow_offset)
+        self.offset_slider.installEventFilter(self)
+        slider_layout.addWidget(self.offset_slider, 0, 0, 1, 3)
+
+        self.min_label = QLabel(str(self.offset_slider.minimum()) + 'px')
+        self.min_label.setFont(self.gui.list_font)
+        slider_layout.addWidget(self.min_label, 1, 0, Qt.AlignmentFlag.AlignLeft)
+
+        self.current_label = QLabel(str(self.offset_slider.value()) + 'px')
+        self.current_label.setFont(self.gui.list_title_font)
+        slider_layout.addWidget(self.current_label, 1, 1, Qt.AlignmentFlag.AlignCenter)
+        self.offset_slider.sliderMoved.connect(lambda value: self.current_label.setText(str(value) + 'px'))
+
+        self.max_label = QLabel(str(self.offset_slider.maximum()) + 'px')
+        self.max_label.setFont(self.gui.list_font)
+        slider_layout.addWidget(self.max_label, 1, 2, Qt.AlignmentFlag.AlignRight)
+
+    def eventFilter(self, obj, evt):
+        if obj == self.offset_slider and evt.type() == QEvent.Type.Wheel:
+            return True
+        elif obj == self.offset_slider and evt.type() == QEvent.Type.MouseButtonRelease:
+            parent = self.parent()
+            while parent.parent():
+                if hasattr(parent, 'mouse_release_signal'):
+                    parent.mouse_release_signal.emit(self.offset_slider.value())
+                    break
+                else:
+                    parent = parent.parent()
+            return super().eventFilter(obj, evt)
         else:
-            bg_color = rgb_color.replace('rgb(', '').replace(')', '')
-        bg_color_split = bg_color.split(', ')
-
-        if len(bg_color_split) == 4:
-            brush = QBrush(
-                QColor(int(bg_color_split[0]), int(bg_color_split[1]), int(bg_color_split[2]), int(bg_color_split[3])))
-        else:
-            brush = QBrush(
-                QColor(int(bg_color_split[0]), int(bg_color_split[1]), int(bg_color_split[2])))
-        brush.setStyle(Qt.BrushStyle.SolidPattern)
-        pen = QPen(Qt.GlobalColor.black)
-        pen.setWidth(2)
-
-        pixmap = QPixmap(48, 48)
-        painter = QPainter(pixmap)
-        painter.setBrush(brush)
-        painter.setPen(pen)
-        painter.setPen(Qt.GlobalColor.black)
-
-        painter.begin(pixmap)
-        painter.fillRect(0, 0, 48, 48, brush)
-        painter.drawRect(QRect(0, 0, 48, 48))
-        painter.end()
-
-        self.setPixmap(pixmap)
-        self.repaint()
-
-    def mouseReleaseEvent(self, evt):
-        super().mouseReleaseEvent(evt)
-        image = self.pixmap().toImage()
-        current_color = image.pixelColor(10, 10)
-        chosen_color = QColorDialog.getColor(current_color, self.gui.main_window, 'Countdown Background Color')
-        rgb_color = f'rgba({chosen_color.red()}, {chosen_color.green()}, {chosen_color.blue()}, {chosen_color.alpha()})'
-        self.make_color_swatch_pixmap(rgb_color)
-        self.color_changed.emit()
+            return super().eventFilter(obj, evt)
 
 
 class PrintDialog(QDialog):
@@ -1933,3 +1810,176 @@ class PrintDialog(QDialog):
         if result == 1:
             printer.setPrinterName(printer_combobox.currentText())
             document.print(printer)
+
+
+class ShadowSlider(QWidget):
+    """
+    Creates a widget containing a QSlider and Label which lets the user set the greyness of the display's shadow
+    :param gui.GUI gui: The current instance of GUI
+    """
+    def __init__(self, gui):
+        """
+        Creates a widget containing a QSlider and Label which lets the user set the greyness of the display's shadow
+        :param gui.GUI gui: The current instance of GUI
+        """
+        super().__init__()
+        self.gui = gui
+        layout = QVBoxLayout()
+        layout.setContentsMargins(0, 0, 0, 0)
+        layout.setSpacing(0)
+        self.setLayout(layout)
+
+        self.color_title = QLabel('Shadow Shade:')
+        self.color_title.setFont(self.gui.list_font)
+        layout.addWidget(self.color_title)
+
+        slider_widget = QWidget()
+        slider_widget.setFixedWidth(300)
+        slider_layout = QGridLayout()
+        slider_layout.setContentsMargins(0, 0, 0, 0)
+        slider_layout.setVerticalSpacing(0)
+        slider_widget.setLayout(slider_layout)
+        layout.addWidget(slider_widget)
+
+        self.color_slider = CustomSlider()
+        self.color_slider.setObjectName('color_slider')
+        self.color_slider.setOrientation(Qt.Orientation.Horizontal)
+        self.color_slider.setFont(self.gui.list_font)
+        self.color_slider.setRange(0, 255)
+        self.color_slider.installEventFilter(self)
+        slider_layout.addWidget(self.color_slider, 0, 0, 1, 3)
+
+        self.min_label = QLabel('Black')
+        self.min_label.setFont(self.gui.list_font)
+        slider_layout.addWidget(self.min_label, 1, 0, Qt.AlignmentFlag.AlignLeft)
+
+        self.color_label = QLabel()
+        color_pixmap = QPixmap(20, 20)
+        color_pixmap.fill(QColor(self.color_slider.value(), self.color_slider.value(), self.color_slider.value()))
+        self.color_label.setPixmap(color_pixmap)
+        self.color_slider.sliderMoved.connect(lambda value: self.change_sample(value))
+        slider_layout.addWidget(self.color_label, 1, 1, Qt.AlignmentFlag.AlignCenter)
+
+        self.max_label = QLabel('White')
+        self.max_label.setFont(self.gui.list_font)
+        slider_layout.addWidget(self.max_label, 1, 2, Qt.AlignmentFlag.AlignRight)
+
+    def change_sample(self, value):
+        new_pixmap = QPixmap(20, 20)
+        new_pixmap.fill(QColor(value, value, value))
+        self.color_label.setPixmap(new_pixmap)
+
+    def eventFilter(self, obj, evt):
+        if obj == self.color_slider and evt.type() == QEvent.Type.Wheel:
+            return True
+        elif obj == self.color_slider and evt.type() == QEvent.Type.MouseButtonRelease:
+            parent = self.parent()
+            while parent.parent():
+                if hasattr(parent, 'mouse_release_signal'):
+                    parent.mouse_release_signal.emit(self.color_slider.value())
+                    break
+                else:
+                    parent = parent.parent()
+            return super().eventFilter(obj, evt)
+        else:
+            return super().eventFilter(obj, evt)
+
+
+class SimpleSplash:
+    """
+    Provides a simple and standardized popup for showing messages
+    """
+
+    def __init__(self, gui, text='', subtitle=False, parent=None):
+        """
+        Provides a simple and standardized popup for showing messages
+        :param gui.GUI gui: the current instance of GUI
+        :param str text: the text to be displayed
+        :param str subtitle: optional: subtitle to be displayed under the text
+        :param obj parent: optional: parent widget for SimpleSplash's main widget
+        """
+        self.gui = gui
+        self.text = text
+
+        self.widget = QWidget(parent)
+        self.widget.setStyleSheet('background: #6060c0;')
+        self.widget.setWindowFlag(Qt.WindowType.FramelessWindowHint)
+        self.widget.setAttribute(Qt.WidgetAttribute.WA_AlwaysStackOnTop)
+        main_layout = QVBoxLayout(self.widget)
+
+        container = QWidget()
+        container.setObjectName('container')
+        container.setStyleSheet(
+            '#container { padding: 30px 20px; border: 3px solid white;}')
+        container.setMinimumWidth(300)
+        layout = QGridLayout(container)
+        main_layout.addWidget(container)
+
+        self.label = QLabel(text)
+        self.label.setFont(self.gui.bold_font)
+        self.label.setStyleSheet('color: white; padding: 10px 5px;')
+        layout.addWidget(self.label, 0, 0, Qt.AlignmentFlag.AlignHCenter)
+
+        if subtitle:
+            self.subtitle_label = QLabel(' ')
+            self.subtitle_label.setFont(self.gui.list_font)
+            self.subtitle_label.setStyleSheet('color: white; padding: 10px 5px;')
+            layout.addWidget(self.subtitle_label, 1, 0, Qt.AlignmentFlag.AlignHCenter)
+
+        container.adjustSize()
+        self.widget.adjustSize()
+        x = int((self.gui.primary_screen.size().width() / 2) - (self.widget.width() / 2))
+        y = int((self.gui.primary_screen.size().height() / 2) - (self.widget.height() / 2))
+        self.widget.move(x, y)
+        self.widget.show()
+        QApplication.processEvents()
+
+
+class StandardItemWidget(QWidget):
+    """
+    Provides a standardized QWidget to be used as a QListWidget ItemWidget
+    """
+    def __init__(self, gui, title, subtitle=None, icon=None, wrap_subtitle=False):
+        super().__init__()
+        self.gui = gui
+        self.setObjectName('item_widget')
+        layout = QHBoxLayout(self)
+
+        self.subtitle = None
+        self.icon = None
+
+        if icon:
+            self.icon = QLabel()
+            self.icon.setAutoFillBackground(False)
+            self.icon.setPixmap(icon)
+            self.icon.adjustSize()
+            layout.addWidget(self.icon)
+
+        text_container = QWidget()
+        text_layout = QVBoxLayout(text_container)
+        layout.addWidget(text_container)
+
+        self.title = QLabel(title)
+        self.title.setAutoFillBackground(False)
+        self.title.setObjectName('lyric_item_widget_title')
+        self.title.setFont(self.gui.list_title_font)
+        self.title.adjustSize()
+        text_layout.addWidget(self.title)
+
+        if subtitle:
+            subtitle = re.sub('<br.*?>', '\n', subtitle)
+            subtitle = re.sub('<.*?>', '', subtitle)
+
+            self.subtitle = QLabel(subtitle)
+            self.subtitle.setAutoFillBackground(False)
+            self.title.setObjectName('lyric_item_widget_text')
+            if wrap_subtitle:
+                self.subtitle.setWordWrap(True)
+            self.subtitle.setFont(self.gui.list_font)
+            self.subtitle.adjustSize()
+            text_layout.addWidget(self.subtitle)
+
+        if not wrap_subtitle:
+            layout.addStretch()
+
+        self.adjustSize()
